% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{activeVersionFile}
\alias{activeVersionFile}
\title{Path to the file recording the active TASSEL version}
\usage{
activeVersionFile()
}
\value{
A character string path.
}
\description{
Path to the file recording the active TASSEL version
}
\keyword{internal}
