% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_tassel_genomic_dataset.R
\name{as.data.frame.TasselGenomicDataset}
\alias{as.data.frame.TasselGenomicDataset}
\title{Coerce a genomic dataset's phenotype data to a data frame}
\usage{
\method{as.data.frame}{TasselGenomicDataset}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{A \code{TasselGenomicDataset} object.}

\item{row.names}{Ignored, present for generic compatibility.}

\item{optional}{Ignored, present for generic compatibility.}

\item{...}{Additional arguments to be passed to or from methods.}
}
\value{
A \code{tibble} of phenotype data.
}
\description{
Returns the joined phenotype data held by a
\code{\linkS4class{TasselGenomicDataset}} as a \code{tibble}.
}
