% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenotypeTableFunctions.R
\name{as.matrix.TasselGenotypePhenotype}
\alias{as.matrix.TasselGenotypePhenotype}
\title{Coerce genotype table to R matrix}
\usage{
\method{as.matrix}{TasselGenotypePhenotype}(tasObj, ...)
}
\arguments{
\item{tasObj}{A \code{TasselGenotypePhenotype} object}

\item{...}{Additional arguments to be passed to or from methods.}
}
\description{
Converts a \code{TasselGenotypePhenotype} class into an R
   matrix if it contains genotype data.
}
