% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_association_results.R
\name{AssociationResults-validity}
\alias{AssociationResults-validity}
\title{AssociationResults validation}
\arguments{
\item{object}{A \code{AssociationResults} object}
}
\description{
Checks if all elements of list are \code{data.frame} objects
}
