% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_association_results_sub.R
\docType{class}
\name{AssociationResultsFast-class}
\alias{AssociationResultsFast-class}
\title{AssociationResultsFast Class}
\description{
Class \code{AssociationResultsFast} defines a \code{rTASSEL}
Class for storing TASSEL 5 FastAssociation GWAS results
}
