% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_association_results_sub.R
\docType{class}
\name{AssociationResultsGLM-class}
\alias{AssociationResultsGLM-class}
\title{AssociationResultsGLM Class}
\description{
Class \code{AssociationResultsGLM} defines a \code{rTASSEL}
Class for storing TASSEL 5 GLM GWAS results
}
