% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_association_results_sub.R
\docType{class}
\name{AssociationResultsMLM-class}
\alias{AssociationResultsMLM-class}
\title{AssociationResultsMLM Class}
\description{
Class \code{AssociationResultsMLM} defines a \code{rTASSEL}
Class for storing TASSEL 5 MLM GWAS results
}
