% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_association_results_sub.R
\docType{class}
\name{AssociationResultsStepwise-class}
\alias{AssociationResultsStepwise-class}
\title{AssociationResultsStepwise Class}
\description{
Class \code{AssociationResultsStepwise} defines a \code{rTASSEL}
Class for storing TASSEL 5 Stepwise results
}
