% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_all_generics.R,
%   R/class_association_results.R
\name{associationType}
\alias{associationType}
\alias{associationType,AssociationResults-method}
\title{Return GWAS association type}
\usage{
associationType(assocRes)

\S4method{associationType}{AssociationResults}(assocRes)
}
\arguments{
\item{assocRes}{a \code{\linkS4class{AssociationResults}} object}
}
\description{
Returns association type for given  \code{\linkS4class{AssociationResults}}
object.
}
