% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_plot_association.R
\name{assocStatsColumnChecker}
\alias{assocStatsColumnChecker}
\title{Sanity checker for association results}
\usage{
assocStatsColumnChecker(x, assocStats, neededCols)
}
\arguments{
\item{x}{A given column to check}

\item{assocStats}{A \code{data.frame} object containing association stats}

\item{neededCols}{A \code{character} vector containing desired columns}
}
\description{
Sanity checker for association results
}
