% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_relatedness_functions.R
\name{asTasselDistanceMatrix}
\alias{asTasselDistanceMatrix}
\title{Coerce matrix to TasselDistanceMatrix object}
\usage{
asTasselDistanceMatrix(m)
}
\arguments{
\item{m}{An object of \code{matrix} class of \eqn{m \times m} structure
(e.g. a pairwise matrix). Additionally, row and column names must be
the same.}
}
\value{
Returns a \code{TasselDistanceMatrix} object.
}
\description{
Coerces an object of \code{matrix} class into an rTASSEL
   object of \code{TasselDistanceMatrix} class.
}
