% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_tassel_genotype.R
\name{chrom}
\alias{chrom}
\title{Select Sites by Chromosome}
\usage{
chrom(...)
}
\arguments{
\item{...}{Character chromosome IDs.}
}
\value{
A \code{\linkS4class{SiteSelector}} object.
}
\description{
Creates a \code{\linkS4class{SiteSelector}} for filtering by one
or more chromosome IDs.
}
\examples{
\dontrun{
gt[, chrom("chr3")]
gt[, chrom("1", "5", "10")]
}

}
