% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassTasselDistanceMatrix.R
\name{colnames,TasselDistanceMatrix-method}
\alias{colnames,TasselDistanceMatrix-method}
\title{Column names}
\usage{
\S4method{colnames}{TasselDistanceMatrix}(x)
}
\arguments{
\item{x}{An object of class \code{TasselDistanceMatrix}.}
}
\description{
Get column names of a \code{TasselDistanceMatrix} object.
}
