% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenotypePhenotypeFunctions.R
\name{combineTasselGenotypePhenotype}
\alias{combineTasselGenotypePhenotype}
\title{Function for combining TASSEL GenotypeTable and Phenotype to GenotypePhenotype class}
\usage{
combineTasselGenotypePhenotype(genotypeTable, phenotype)
}
\arguments{
\item{genotypeTable}{Java object of GenotypeTable}

\item{phenotype}{Java object of Phenotype}
}
\value{
a TASSEL GenotypePhenotype object
}
\description{
Function for combining TASSEL GenotypeTable and Phenotype to GenotypePhenotype class
}
