% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PluginSupport.R
\name{createTasselDataSet}
\alias{createTasselDataSet}
\title{Method for wrapping TASSEL objects in Datum and Dataset}
\usage{
createTasselDataSet(tasObj)
}
\arguments{
\item{tasObj}{a vector of TASSEL objects}
}
\value{
a TASSEL Dataset
}
\description{
Method for wrapping TASSEL objects in Datum and Dataset
}
