% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassTasselDistanceMatrix.R
\name{dim,TasselDistanceMatrix-method}
\alias{dim,TasselDistanceMatrix-method}
\title{Get dimensions of TasselDistanceMatrix object}
\usage{
\S4method{dim}{TasselDistanceMatrix}(x)
}
\description{
Retrieves dimensions of a \code{TasselDistanceMatrix} object
   as a vector (e.g. \code{c(10, 10)}).
}
