% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_relatedness_functions.R
\name{distanceMatrix}
\alias{distanceMatrix}
\title{Create a TASSEL distance matrix}
\usage{
distanceMatrix(tasObj)
}
\arguments{
\item{tasObj}{An object of class \code{\linkS4class{TasselGenotype}} or
\code{\linkS4class{TasselGenomicDataset}}. Objects of the deprecated
\code{TasselGenotypePhenotype} class are still accepted.}
}
\value{
Returns a `TasselDistanceMatrix` object.
}
\description{
This function will calculate a distance matrix using
   TASSEL's \code{DistanceMatrix} plugin.
}
