% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TasselPluginWrappers.R
\name{distanceMatrixPlugin}
\alias{distanceMatrixPlugin}
\title{Distance Matrix Plugin}
\usage{
distanceMatrixPlugin(genotypeTable)
}
\arguments{
\item{genotypeTable}{an rTASSEL \code{TasselGenotypePhenotype} object}
}
\description{
Distance Matrix Plugin
}
