% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysisRelatednessFunctions.R
\name{distanceToRMatrix}
\alias{distanceToRMatrix}
\title{Convert TASSEL distance matrix object to an R matrix class}
\usage{
distanceToRMatrix(distJobj)
}
\arguments{
\item{distJobj}{A TASSEL distance matrix object.}
}
\value{
Returns an R \code{matrix} object.
}
\description{
This function will take a TASSEL distance matrix object and
   convert it into an R \code{matrix} object.
}
