% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PullFunctions.R
\name{.dataSetToListOfRObjects}
\alias{.dataSetToListOfRObjects}
\title{Converts TASSEL dataset to List of R objects - either DataFrames or TasselGenotypePhenotype S4 Class}
\usage{
.dataSetToListOfRObjects(jtsDataSet)
}
\description{
Converts TASSEL dataset to List of R objects - either DataFrames or TasselGenotypePhenotype S4 Class
}
