% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_qq.R
\name{expectedQuantiles}
\alias{expectedQuantiles}
\title{Get expected quantiles from p-val vectors}
\usage{
expectedQuantiles(pVal, trait)
}
\arguments{
\item{pVal}{A \code{numeric} vector containing association p-values}

\item{trait}{A trait ID that correlates with p-value data}
}
\description{
Get expected quantiles from p-val vectors
}
