% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{FactorTable-class}
\alias{FactorTable-class}
\alias{show,FactorTable-method}
\title{FactorTable Class}
\usage{
\S4method{show}{FactorTable}(object)
}
\arguments{
\item{object}{a \code{FactorTable} class object}
}
\description{
Class \code{FactorTable} defines a \code{rTASSEL}
   Class for storing TASSEL 6 \code{FactorTable} objects.

Prints out information related taxa, positions, genotype, and
   phenotype information.
}
