% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FactorTableFunctions.R
\name{factorTableDF}
\alias{factorTableDF}
\title{Get haplotype ID data frame}
\usage{
factorTableDF(factorTable)
}
\arguments{
\item{factorTable}{An object of clase \code{FactorTable}.}
}
\value{
Returns a \code{data.frame} of haplotype and taxa IDs.
}
\description{
Returns a \code{data.frame} of haplotype and taxa IDs from a
   \code{FactorTable} object.
}
