% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{FeatureTable-class}
\alias{FeatureTable-class}
\alias{show,FeatureTable-method}
\title{FeatureTable Class}
\usage{
\S4method{show}{FeatureTable}(object)
}
\arguments{
\item{object}{a \code{FeatureTable} class object}
}
\description{
Class \code{FeatureTable} defines a \code{rTASSEL}
   Class for storing TASSEL 6 \code{FeatureTable} objects.

Prints out information related taxa, positions, genotype, and
   phenotype information.
}
