% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureTableFunctions.R
\name{featureTableDF}
\alias{featureTableDF}
\title{Get haplotype ID data frame}
\usage{
featureTableDF(featureTable)
}
\arguments{
\item{featureTable}{An object of clase \code{FeatureTable}.}
}
\value{
Returns a \code{data.frame} of haplotype and taxa IDs.
}
\description{
Returns a \code{data.frame} of haplotype and taxa IDs from a
   \code{FeatureTable} object.
}
