% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PullFunctions.R
\name{genomicRanges}
\alias{genomicRanges}
\title{Constructor for GRanges (GenomicRanges) class object}
\usage{
genomicRanges(jtsGenoTable)
}
\description{
Constructor for GRanges (GenomicRanges) class object
}
