% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_all_generics.R,
%   R/class_tassel_genomic_dataset.R
\name{genotype}
\alias{genotype}
\alias{genotype,TasselGenomicDataset-method}
\title{Return the genotype component of an rTASSEL object}
\usage{
genotype(object, ...)

\S4method{genotype}{TasselGenomicDataset}(object)
}
\arguments{
\item{object}{an \code{rTASSEL} object}

\item{...}{Additional arguments, for use in specific methods}
}
\description{
Returns the genotype data held by a composite \code{rTASSEL} object, such as
the \code{\linkS4class{TasselGenotype}} carried by a
\code{\linkS4class{TasselGenomicDataset}}.
}
