% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{GenotypeTable-class}
\alias{GenotypeTable-class}
\title{Class GenotypeTable}
\description{
Class GenotypeTable
}
\details{
Class \code{GenotypeTable} defines a GenotypeTable from TASSEL
}
