% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{GenotypeTable-class}
\alias{GenotypeTable-class}
\alias{GenotypeTable}
\title{Constructor for GenotypeTable class object}
\usage{
GenotypeTable(path)
}
\description{
Constructor for GenotypeTable class object
}
