% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype_functions.R
\name{getPhenotypeDF}
\alias{getPhenotypeDF}
\title{Get a phenotype data frame from a TASSEL object}
\usage{
getPhenotypeDF(tasObj)
}
\arguments{
\item{tasObj}{An object of class \code{\linkS4class{TasselPhenotype}} or
\code{\linkS4class{TasselGenomicDataset}}. Objects of the deprecated
\code{TasselGenotypePhenotype} class are still accepted.}
}
\value{
A \code{tibble}, one row per observation.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function will extract phenotype data from an object that contains
   it. Use \code{\link[base]{as.data.frame}()} on a
   \code{\linkS4class{TasselPhenotype}} or
   \code{\linkS4class{TasselGenomicDataset}} instead.
}
\seealso{
\code{\link[base]{as.data.frame}}
}
