% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureTableFunctions.R
\name{getSumExp}
\alias{getSumExp}
\title{Create Summarized Experiment from a TASSEL Genotype Table}
\usage{
getSumExp(featureTable)
}
\arguments{
\item{featureTable}{An object of clase \code{FeatureTable}.}
}
\value{
Returns a \code{SummarizedExperiment} of TASSEL genotype data.
}
\description{
This function will generate an object of
   \code{SummarizedExperiment} class for marker data derived from a
   \code{FeatureTable} class object.
}
