% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{getTASSELJarName}
\alias{getTASSELJarName}
\title{Get the TASSEL JAR filename for a given version}
\usage{
getTASSELJarName(version = TASSEL_MAVEN$VERSION)
}
\arguments{
\item{version}{TASSEL version string.}
}
\value{
A character string JAR filename.
}
\description{
Get the TASSEL JAR filename for a given version
}
\keyword{internal}
