% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{getTASSELJavaDir}
\alias{getTASSELJavaDir}
\title{Get the root directory holding all cached TASSEL versions}
\usage{
getTASSELJavaDir()
}
\value{
A character string path.
}
\description{
Get the root directory holding all cached TASSEL versions
}
\keyword{internal}
