% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{getTASSELThinJarName}
\alias{getTASSELThinJarName}
\title{Get the JAR filename for a thin (dependency-less) TASSEL release}
\usage{
getTASSELThinJarName(version)
}
\arguments{
\item{version}{TASSEL version string.}
}
\value{
A character string JAR filename.
}
\description{
Get the JAR filename for a thin (dependency-less) TASSEL release
}
\keyword{internal}
