% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{installSourceFile}
\alias{installSourceFile}
\title{Path to the file recording where a cached version came from}
\usage{
installSourceFile(version)
}
\arguments{
\item{version}{TASSEL version string.}
}
\value{
A character string path.
}
\description{
Path to the file recording where a cached version came from
}
\keyword{internal}
