% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_all_generics.R,
%   R/class_tassel_distance_matrix.R, R/class_tassel_phenotype.R,
%   R/class_tassel_genotype.R, R/class_tassel_genomic_dataset.R
\name{javaRefObj}
\alias{javaRefObj}
\alias{javaRefObj,TasselDistanceMatrix-method}
\alias{javaRefObj,TasselPhenotype-method}
\alias{javaRefObj,TasselGenotype-method}
\alias{javaRefObj,TasselGenomicDataset-method}
\title{Return \code{rJava} reference object}
\usage{
javaRefObj(object, ...)

\S4method{javaRefObj}{TasselDistanceMatrix}(object)

\S4method{javaRefObj}{TasselPhenotype}(object)

\S4method{javaRefObj}{TasselGenotype}(object)

\S4method{javaRefObj}{TasselGenomicDataset}(object)
}
\arguments{
\item{object}{an \code{rTASSEL} object}

\item{...}{Additional arguments, for use in specific methods}
}
\description{
Returns the \code{rJava} memory reference for a given \code{rTASSEL} object
}
