% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_relatedness_functions.R
\name{kinshipMatrix}
\alias{kinshipMatrix}
\title{Create a TASSEL kinship matrix}
\usage{
kinshipMatrix(
  tasObj,
  method = "Centered_IBS",
  maxAlleles = 6,
  algorithmVariation = "Observed_Allele_Freq"
)
}
\arguments{
\item{tasObj}{An object of class \code{\linkS4class{TasselGenotype}} or
\code{\linkS4class{TasselGenomicDataset}}. Objects of the deprecated
\code{TasselGenotypePhenotype} class are still accepted.}

\item{method}{A Kinship method. Defaults to \code{Centered_IBS}. Other
options include \code{Normalized_IBS}, \code{Dominance_Centered_IBS},
and \code{Dominance_Normalized_IBS}.}

\item{maxAlleles}{Maximum number of alleles. Can be within the range of
\code{2} to \code{6}.}

\item{algorithmVariation}{Algorithm variation. If
\code{Dominance_Centered_IBS} is selected, users can switch between
\code{Observed_Allele_Freq} and \code{Proportion_Heterozygous}.}
}
\value{
Returns a `TasselDistanceMatrix` object.
}
\description{
This function will calculate a kinship matrix using
   TASSEL's \code{Kinship} plugin and respective parameters.
}
