% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TasselPluginWrappers.R
\name{kinshipPlugin}
\alias{kinshipPlugin}
\title{Kinship Plugin}
\usage{
kinshipPlugin(genotypeTable, method = "Centered_IBS", maxAlleles = 6,
  algorithmVariation = "Observed_Allele_Freq")
}
\arguments{
\item{genotypeTable}{an rTASSEL \code{TasselGenotypePhenotype} object}

\item{method}{a Kinship method}

\item{maxAlleles}{Maximum number of alleles}

\item{algorithmVariation}{Algorithm variation}
}
\description{
Kinship Plugin
}
