% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KinshipFunctions.R
\name{kinshipRMatrix}
\alias{kinshipRMatrix}
\title{Convert TASSEL kinship object to an R matrix class}
\usage{
kinshipRMatrix(kinJobj)
}
\arguments{
\item{kinJobj}{a TASSEL kinship object}
}
\description{
Convert TASSEL kinship object to an R matrix class
}
