% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysisRelatednessFunctions.R
\name{kinshipToRMatrix}
\alias{kinshipToRMatrix}
\title{Convert TASSEL kinship matrix object to an R matrix class}
\usage{
kinshipToRMatrix(kinJobj)
}
\arguments{
\item{kinJobj}{A TASSEL kinship object.}
}
\value{
Returns an R \code{matrix} object.
}
\description{
This function will take a TASSEL kinship object and convert
   it into an R \code{matrix} object.
}
