% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysisDiveristyVisFunctions.R
\name{ldApp}
\alias{ldApp}
\title{LD visualization application}
\usage{
ldApp(ldData)
}
\arguments{
\item{ldData}{An LD data frame.}
}
\description{
Runs an interactive visualizer for an LD object
}
