% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ld_results.R
\name{LDResults-validity}
\alias{LDResults-validity}
\title{LDResults validation}
\arguments{
\item{object}{An \code{LDResults} object.}
}
\description{
Validates that an \code{LDResults} object has a properly formed
\code{results} data frame with the required LD columns and that
\code{ldType} and \code{hetCalls} contain allowed values.
}
