% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicPrediction.R
\name{leaveOneFamilyOut}
\alias{leaveOneFamilyOut}
\title{Leave one family out cross-validation}
\usage{
leaveOneFamilyOut(phenoFamilyDF, tasKin)
}
\arguments{
\item{phenoFamilyDF}{An object of class \code{TasselGenotypePenotype} that
contains a phenotype object with family data.}

\item{tasKin}{A TASSEL kinship object.}
}
\description{
Cross-validation by leaving one family out.
}
