% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_diversity_functions.R
\name{linkageDiseq}
\alias{linkageDiseq}
\title{Calculate linkage disequilibrium from an rTASSEL genotype
  dataset.}
\usage{
linkageDiseq(
  tasObj,
  ldType = c("SlidingWindow", "All"),
  windowSize = NULL,
  hetCalls = c("missing", "ignore", "third"),
  verbose = TRUE
)
}
\arguments{
\item{tasObj}{An object of class \code{\linkS4class{TasselGenotype}} or
\code{\linkS4class{TasselGenomicDataset}}. Objects of the deprecated
\code{TasselGenotypePhenotype} class are still accepted.}

\item{ldType}{How do you want LD calculated? Currently, the available
options are \code{"All"} and \code{"SlidingWindow"}. If
\code{All} is selected, LD will be calculated for every
combination of sites in the alignment (NOTE: this may produce a
massive series of combinations; use only on heavily filtered
genotype tables). If \code{SlidingWindow} is selected, LD will
be calculated for sites within a window of sites surrounding the
current site.}

\item{windowSize}{What size do you want your LD analysis window? If you
have chosen \code{SlidingWindow} for the \code{ldType} parameter, you
will need to specify window size.}

\item{hetCalls}{How should heterozygous calls be handled? Current options
are \code{"ignore"} (ignore altogether), \code{"missing"}
(set to missing), and \code{"third"} (treat as third state).}

\item{verbose}{Display messages? Defaults to \code{TRUE}.}
}
\value{
An object of class \code{\linkS4class{LDResults}}.
}
\description{
Generates a linkage disequilibrium (LD) data set from SNP data.
}
\details{
Linkage disequilibrium between any set of polymorphisms can be
  estimated by initially filtering a genotype dataset and then using
  this function. At this time, \eqn{D'}, \eqn{r^{2}} and P-values will be estimated. The
  current version calculates LD between haplotypes with known phase only
  (unphased diploid genotypes are not supported; see PowerMarker or
  Arlequin for genotype support).
  \itemize{
    \item \eqn{D'} is the standardized disequilibrium coefficient, a useful
    statistic for determining whether recombination or homoplasy has
    occurred between a pair of alleles.
    \item \eqn{r^{2}} represents the correlation between alleles at two loci, which
    is informative for evaluating the resolution of association approaches.
  }
  \eqn{D'} and \eqn{r^{2}} can be calculated when only two alleles are present. If more
  than two alleles, only the two most frequent alleles are used. P-values
  are determined by a two-sided Fisher's Exact test is calculated. Since LD
  is meaningless when scored with very small sample sizes, a minimum of 20
  taxa must be present to calculate LD and there must be 2 or more minor
  alleles.
}
\seealso{
\code{\link{linkageDiseq}}, \code{\link{plotLD}}
}
