% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDApp.R
\name{linkageDiseqApp}
\alias{linkageDiseqApp}
\title{LD visualization application}
\usage{
linkageDiseqApp(ldData)
}
\arguments{
\item{ldData}{An LD data frame.}
}
\description{
Runs an interactive visualizer for an LD object
}
