% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysisAssociationVisFunctions.R
\name{manhattanPlot}
\alias{manhattanPlot}
\title{Create Manhattan plot from rTASSEL association output}
\usage{
manhattanPlot(assocStats, trait, threshold, colors = c("#91baff",
  "#3e619b"), showSigMarkers = FALSE)
}
\arguments{
\item{assocStats}{rTASSEL association statistical output. This output is
derived from association output dat frames ending in \code{_Stats}.
This data contains the necessary column variables for this plot to work.}

\item{trait}{Which phenotypic trait do you want to plot?}

\item{threshold}{User-defined \eqn{p}-value threshold for significant
marker determination. Once specified any marker points higher than this
line will be highlighted.}

\item{colors}{A vector of colors used for differentiating multiple
chromosomes. Defaults to 2 shades of blue.}

\item{showSigMarkers}{Should significantly highlighted markers be labelled?
If \code{TRUE}, marker points will labelled with their respective
marker labels. Defaults to \code{FALSE}.}
}
\value{
Returns a \code{ggplot2} object
}
\description{
This function allows for quick generation of a Manhattan
   plot from rTASSEL association statistical output data.
}
