% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_mds_results.R
\docType{class}
\name{MDSResults-class}
\alias{MDSResults-class}
\alias{show,MDSResults-method}
\title{MDSResults Class}
\usage{
\S4method{show}{MDSResults}(object)
}
\arguments{
\item{object}{a \code{\linkS4class{MDSResults}} object}
}
\description{
Class \code{MDSResults} defines a \code{rTASSEL}
Class for storing TASSEL 5 MDS results

Prints outs summary information from \code{MDSResults} objects
}
\section{Slots}{

\describe{
\item{\code{results}}{A named list of \code{tibble} objects containing summary
results}

\item{\code{jObj}}{An rJava reference object pointing to MDS results in Java memory}
}}

