% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_mds_results.R
\name{MDSResults-validity}
\alias{MDSResults-validity}
\title{MDSResults validation}
\arguments{
\item{object}{A \code{MDSResults} object}
}
\description{
Checks if all elements of list are \code{data.frame} objects
}
