% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassTasselDistanceMatrix.R
\name{nrow,TasselDistanceMatrix-method}
\alias{nrow,TasselDistanceMatrix-method}
\title{Number of rows}
\usage{
\S4method{nrow}{TasselDistanceMatrix}(x)
}
\arguments{
\item{x}{An object of class \code{TasselDistanceMatrix}.}
}
\description{
Get number of rows of a \code{TasselDistanceMatrix} object.
}
