% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_pca_results.R
\name{PCAResults-validity}
\alias{PCAResults-validity}
\title{PCAResults validation}
\arguments{
\item{object}{A \code{PCAResults} object}
}
\description{
Checks if all elements of list are \code{data.frame} objects
}
