% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_all_generics.R,
%   R/class_tassel_genomic_dataset.R
\name{phenotype}
\alias{phenotype}
\alias{phenotype,TasselGenomicDataset-method}
\title{Return the phenotype component of an rTASSEL object}
\usage{
phenotype(object, ...)

\S4method{phenotype}{TasselGenomicDataset}(object)
}
\arguments{
\item{object}{an \code{rTASSEL} object}

\item{...}{Additional arguments, for use in specific methods}
}
\description{
Returns the phenotype data held by a composite \code{rTASSEL} object, such as
the \code{\linkS4class{TasselPhenotype}} carried by a
\code{\linkS4class{TasselGenomicDataset}}.
}
