% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_manhattan.R
\name{plotManhattan}
\alias{plotManhattan}
\title{Create Manhattan plot from rTASSEL association output}
\usage{
plotManhattan(
  assocRes,
  trait = NULL,
  threshold = NULL,
  colors = c("#91baff", "#3e619b"),
  pltTheme = c("default", "classic")
)
}
\arguments{
\item{assocRes}{An object of type \code{AssociationResults}}

\item{trait}{Which phenotypic trait do you want to plot? If set to
\code{NULL}, this will generate a faceted plot with all mapped traits}

\item{threshold}{User-defined \eqn{-log_{10}(p)}-value threshold for
significant marker determination. Once specified any marker points
higher than this line will be highlighted.}

\item{colors}{A vector of \code{character} colors used for differentiating
multiple chromosomes. Defaults to 2 shades of blue.}

\item{pltTheme}{What theme would like to display for the plot? Only
supports one theme currently.}
}
\value{
Returns a \code{ggplot2} object
}
\description{
This function allows for quick generation of a Manhattan
   plot from rTASSEL association statistical output data.
}
