% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_manhattan.R
\name{plotManhattanCore}
\alias{plotManhattanCore}
\title{Core visual engine for Manhattan plotting}
\usage{
plotManhattanCore(params)
}
\arguments{
\item{params}{A list of parameter variables}
}
\description{
Core visual engine for Manhattan plotting
}
