% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_manhattan_qc.R
\name{plotManhattanQCCore}
\alias{plotManhattanQCCore}
\title{Core visual engine for QQ plotting}
\usage{
plotManhattanQCCore(params)
}
\arguments{
\item{params}{A list of parameter variables}
}
\description{
Core visual engine for QQ plotting
}
