% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pca.R
\name{plotPCA}
\alias{plotPCA}
\title{Generate PCA plot}
\usage{
plotPCA(
  pcaObj,
  x = 1,
  y = 2,
  cluster = FALSE,
  nClust = 2,
  metadata = NULL,
  mCol = NULL,
  interactive = FALSE,
  pointColor = "black"
)
}
\arguments{
\item{pcaObj}{A \code{PCAResults} object containing eigenvalue data (
e.g \code{PC_Datum}).}

\item{x}{Principal component number to plot on x-axis.}

\item{y}{Principal component number to plot on y-axis.}

\item{cluster}{Plot points by hierachical clustering groups? Defaults to
\code{FALSE}.}

\item{nClust}{Number of clusters to report.}

\item{metadata}{A \code{data.frame} object of additional categorical
information for each sample/taxa. Defaults to \code{NULL}.}

\item{mCol}{What metadata column do you want to plot? Needs metadata
object to work.}

\item{interactive}{Do you want to produce an interactive visualization?
Defaults to \code{FALSE}.}

\item{pointColor}{color of non-categorical PCA data points.}
}
\description{
This function will generate a general 2D scatterplot for a given set
of principal components.
}
