% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_qq.R
\name{plotQQCore}
\alias{plotQQCore}
\title{Core visual engine for QQ plotting}
\usage{
plotQQCore(params)
}
\arguments{
\item{params}{A list of parameter variables}
}
\description{
Core visual engine for QQ plotting
}
