% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scree.R
\name{plotScree}
\alias{plotScree}
\title{Generate scree plots}
\usage{
plotScree(
  pcaObj,
  nComp = 10,
  interactive = FALSE,
  lineColor = "grey",
  pointColor = "black"
)
}
\arguments{
\item{pcaObj}{A \code{PCAResults} object containing eigenvalue data (
e.g \code{Eigenvalues_Datum}).}

\item{nComp}{Total number of principal components to plot. Defaults to
\code{10}.}

\item{interactive}{Do you want to produce an interactive visualization?
Defaults to \code{FALSE}.}

\item{lineColor}{Color of scree line.}

\item{pointColor}{Color of scree points.}
}
\description{
This function will generate a line plot of eigenvalues of PCs. These
plots can be used to determine the number of factors to retain for
downstream analyses (e.g GWAS).
}
