% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{PositionList-class}
\alias{PositionList-class}
\title{A R Wrapper for the PositionList class}
\description{
A R Wrapper for the PositionList class
}
\details{
Class \code{PositionList} defines a PositionList from TASSEL
}
