% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_all_generics.R,
%   R/class_tassel_genotype.R, R/class_tassel_genomic_dataset.R,
%   R/position_list_functions.R
\name{positionList}
\alias{positionList}
\alias{positionList,TasselGenotype-method}
\alias{positionList,TasselGenomicDataset-method}
\alias{positionList,TasselGenotypePhenotype-method}
\title{Get position list metadata from genotype data}
\usage{
positionList(tasObj)

\S4method{positionList}{TasselGenotype}(tasObj)

\S4method{positionList}{TasselGenomicDataset}(tasObj)

\S4method{positionList}{TasselGenotypePhenotype}(tasObj)
}
\arguments{
\item{tasObj}{A \code{TasselGenotypePhenotype} or \code{TasselGenotype}
object containing genotype data.}
}
\value{
A \code{tibble} of positional metadata.
}
\description{
Returns positional metadata (chromosome, position, etc.) from a
\code{\linkS4class{TasselGenotypePhenotype}} or
\code{\linkS4class{TasselGenotype}} object.
}
