% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\docType{methods}
\name{positions}
\alias{positions}
\alias{positions,GenotypeTable-method}
\title{Get positions from TASSEL objects}
\usage{
positions(object)

\S4method{positions}{GenotypeTable}(object)
}
\description{
Get positions from TASSEL objects

Get positions for GenotypeTable class objects
}
