% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllClasses.R
\docType{methods}
\name{positions}
\alias{positions}
\title{Get positions for GenotypeTable class objects}
\usage{
positions(object)

\S4method{positions}{GenotypeTable}(object)
}
\description{
Get positions for GenotypeTable class objects
}
