% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_manhattan.R
\name{primeManhattanData}
\alias{primeManhattanData}
\title{plotManhattan dataframe primer}
\usage{
primeManhattanData(params)
}
\arguments{
\item{params}{A list of parameter variables}
}
\description{
plotManhattan dataframe primer
}
