% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAmpSeq.R
\name{rAmpSeq}
\alias{rAmpSeq}
\title{Run rAmpSeq in R}
\usage{
rAmpSeq(
  refGenomFile,
  kmerLength,
  minAmpLength,
  maxAmpLenght,
  maxPrimerObs,
  minCopyNum,
  maxCopyNum,
  ambigBPs = FALSE
)
}
\arguments{
\item{kmerLength}{Length of k-mer to be identified.}

\item{minAmpLength}{Minimum length of amplicons to be produced.}

\item{maxPrimerObs}{Maximum number of "good" primers to be selected.}

\item{minCopyNum}{Minimum amplicon copy threshold.}

\item{maxCopyNum}{Maximum amplicon copy threshold.}

\item{ambigBPs}{Do you want ambiguous basepairs to be generated? Defaults
to \code{FALSE}. See details for further explanation.}

\item{refGenomeFile}{A reference genome file (e.g. \code{*.fa.gz}).}

\item{maxAmpLength}{Maximum length of amplicons to be produced.}
}
\description{
Runs the rAmpSeq pipeline in R with user defined methods.
}
\details{
Currently, the \code{ambigBPs} parameters convert every third
  base pair to an \code{A}. For example, if a k-mer of length 12 was
  generated:

  \code{A C C G T A T T C G A C}

  would be converted to:

  \code{A C \strong{A} G T \strong{A} T T \strong{A} G A \strong{A}}
}
