% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PluginSupport.R
\name{randOrFixed}
\alias{randOrFixed}
\title{Method for wrapping TASSEL objects in Datum and Dataset}
\usage{
randOrFixed()
}
\arguments{
\item{...}{a collection of TASSEL objects}
}
\value{
a TASSEL Dataset
}
\description{
Method for wrapping TASSEL objects in Datum and Dataset
}
