% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysisRelatednessFunctions.R
\name{readDistanceMatrix}
\alias{readDistanceMatrix}
\title{read TASSEL distance matrix object from file}
\usage{
readDistanceMatrix(file)
}
\arguments{
\item{file}{A file path of type \code{character}}
}
\value{
Returns a \code{TasselDistanceMatrix} object.
}
\description{
This function will read a TASSEL distance matrix from
   file and convert it into a \code{TasselDistanceMatrix} object.
}
