% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FactorTableFunctions.R
\name{readFactorTable}
\alias{readFactorTable}
\title{Wrapper function of FactorTable class for haplotype data data}
\usage{
readFactorTable(file, verbose = TRUE)
}
\arguments{
\item{file}{A haplotype or factor data path (e.g. \code{*.vcf}).}

\item{verbose}{Show messages to console? Defaults to \code{TRUE}.}
}
\value{
Returns an object of \code{FactorTable} class.
}
\description{
This function is a wrapper for the
   \code{FactorTable} class. It is used for storing genotype
   information into a class object.
}
