% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureTableFunctions.R
\name{readFeatureTableFromBrapi}
\alias{readFeatureTableFromBrapi}
\title{Read feature information from a BrAPI web service}
\usage{
readFeatureTableFromBrapi(brapiObj, method, verbose = TRUE)
}
\arguments{
\item{brapiObj}{An BrAPI connection object of \code{BrapiCon} class.}

\item{method}{A BrAPI method ID, currently linked to PHG methods.}

\item{verbose}{Show messages to console? Defaults to \code{TRUE}.}
}
\value{
Returns an object of \code{FeatureTable} class.
}
\description{
Returns PHG feature information hosted on a BrAPI server.
}
