% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureTableFunctions.R
\name{readFeatureTableFromVCF}
\alias{readFeatureTableFromVCF}
\title{Read feature information from \code{HVCF} files}
\usage{
readFeatureTableFromVCF(file, verbose = TRUE)
}
\arguments{
\item{file}{A haplotype or feature data path (e.g. \code{*.vcf}).}

\item{verbose}{Show messages to console? Defaults to \code{TRUE}.}
}
\value{
Returns an object of \code{FeatureTable} class.
}
\description{
This function is a wrapper for the
   \code{FeatureTable} class. It is used for storing genotype feature
   information into a class object.
}
