% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_tassel_genotype.R
\name{readGenotype}
\alias{readGenotype}
\title{Helper function to build a TasselGenotype object}
\usage{
readGenotype(x, sortPositions = FALSE, keepDepth = FALSE)
}
\description{
Helper function to build a TasselGenotype object
}
