% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotype_phenotype_function.R
\name{readGenotypePhenotype}
\alias{readGenotypePhenotype}
\title{Wrapper function of TasselGenotypePhenotype class for
   GenotypePhenotype combined data}
\usage{
readGenotypePhenotype(genoPathOrObj, phenoPathDFOrObj, ...)
}
\arguments{
\item{genoPathOrObj}{a path to a genotype file (e.g. VCF, hmp, etc.) or
TASSEL Genotype Obj}

\item{phenoPathDFOrObj}{a path, a data frame of phenotypic data, or TASSEL
Phenotype Obj}

\item{...}{Additional parameters to be sent to the function. Currently,
if an R data frame object is passed, additional parameters will also
need to be entered for this process. These parameters are derived from
the \code{readPhenotypeFromDataFrame} function. Mainly, \code{taxaID}
is required. If you would like to specify depth retention and position
sorting in Genotype Tables from a path, indicate them here. See
\code{readGenotypeTableFromPath()} for more detail.}
}
\value{
Returns an object of \code{TasselGenotypePhenotype} class.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Creates a Java GenotypePhenotype object which is used for
   the deprecated \code{TasselGenotypePhenotype} object construction. The
   Java GenotypePhenotype object is created via an \code{intersect} method
   from TASSEL. Use \code{\link{readGenomicDataset}()} instead, which
   returns a \code{\linkS4class{TasselGenomicDataset}} object.
}
\seealso{
\code{\link{readGenomicDataset}}
}
