% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenotypeTableFunctions.R
\name{readGenotypeTable}
\alias{readGenotypeTable}
\title{Wrapper function of TasselGenotypePhenotype class for genotype
   data}
\usage{
readGenotypeTable(path)
}
\arguments{
\item{path}{a genotype data path (e.g. \code{*.VCF, *.hmp}, etc.)}
}
\description{
This function is a wrapper for the 
   \code{TasselGenotypePhenotype} class. It is used for storing genotype
   information into a class object.
}
