% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenotypeTableFunctions.R
\name{readGenotypeTableFromBrapi}
\alias{readGenotypeTableFromBrapi}
\title{Read \code{GenotypeTable} objects from BrAPI}
\usage{
readGenotypeTableFromBrapi(brapiObj, variantSetDbId = NULL)
}
\arguments{
\item{brapiObj}{A \code{BrapiCon} object.}

\item{variantSetDbId}{What set of variants do you want to read from BrAPI?}
}
\value{
Returns an object of \code{TasselGenotypePhenotype} class.
}
\description{
Reads \code{GenotypeTable} objects from BrAPI
}
