% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotype_table_functions.R
\name{readGenotypeTableFromGigwa}
\alias{readGenotypeTableFromGigwa}
\title{Read genotype data from GIGWA using QBMS}
\usage{
readGenotypeTableFromGigwa(gigwa)
}
\arguments{
\item{gigwa}{A \code{QBMS}-formatted GIGWA data frame object}
}
\value{
A \code{TasselGenotype} object.
}
\description{
Reads and stores genotype information from a
   \code{QBMS}-formatted data frame from a GIGWA server
}
