% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotype_table_functions.R
\name{readGenotypeTableFromPath}
\alias{readGenotypeTableFromPath}
\title{Wrapper function of TasselGenotypePhenotype class for genotype
   data}
\usage{
readGenotypeTableFromPath(path, keepDepth = FALSE, sortPositions = FALSE)
}
\arguments{
\item{path}{A genotype data path (e.g. \code{*.VCF, *.hmp}, etc.).}

\item{keepDepth}{Should depth be kept? Defaults to \code{FALSE}.}

\item{sortPositions}{Should positions be sorted? Defaults to \code{FALSE}.}
}
\value{
Returns an object of \code{TasselGenotypePhenotype} class.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is a wrapper for the deprecated
\code{TasselGenotypePhenotype} class. It is used for storing genotype
information into a class object. Use \code{\link{readGenotype}()} instead,
which returns a \code{\linkS4class{TasselGenotype}} object.
}
\seealso{
\code{\link{readGenotype}}
}
