% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhenotypeFunctions.R
\name{readPhenotypeFromBrapi}
\alias{readPhenotypeFromBrapi}
\title{Read phenotype information from a BrAPI web service}
\usage{
readPhenotypeFromBrapi(brapiObj, studyName = NULL, verbose = TRUE)
}
\arguments{
\item{brapiObj}{An BrAPI connection object of \code{BrapiCon} class.}

\item{studyName}{Return a specified study. If \code{NULL}, all data will be
returned. Defaults to \code{NULL}.}

\item{verbose}{Show messages to console? Defaults to \code{TRUE}.}
}
\value{
Returns an object of \code{data.frame} class.
}
\description{
Reads phenotype information from BreedBase using a BrAPI rest
   client.
}
