% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype_functions.R
\name{readPhenotypeFromPath}
\alias{readPhenotypeFromPath}
\title{Wrapper function of TasselGenotypePhenotype class for phenotype
   data from a path.}
\usage{
readPhenotypeFromPath(path)
}
\arguments{
\item{path}{A phenotype data path.}
}
\value{
Returns an object of \code{TasselGenotypePhenotype} class.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is a wrapper for the deprecated
\code{TasselGenotypePhenotype} class. It is used for storing phenotype
information into a class object. This will read in phenotype data from
a path. Use \code{\link{readPhenotype}()} instead, which returns a
\code{\linkS4class{TasselPhenotype}} object.
}
\seealso{
\code{\link{readPhenotype}}
}
