% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{readPhenotypeTable}
\alias{readPhenotypeTable}
\title{Wrapper function of TasselGenotypePhenotype class for phenotype
   data}
\usage{
readPhenotypeTable(path)
}
\arguments{
\item{path}{a phenotype data path or \code{R} data frame}
}
\description{
This function is a wrapper for the 
   \code{TasselGenotypePhenotype} class. It is used for storing phenotype
   information into a class object.
}
