% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureTableFunctions.R
\name{refRanges}
\alias{refRanges}
\title{Get reference ranges}
\usage{
refRanges(featureTable)
}
\arguments{
\item{featureTable}{An object of clase \code{FeatureTable}.}
}
\value{
Returns a \code{GRanges} object of reference range data.
}
\description{
Returns reference range intervals from a \code{FeatureTable}
   object.
}
