% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_all_generics.R,
%   R/class_association_results.R, R/class_ld_results.R, R/class_mds_results.R,
%   R/class_pca_results.R
\name{reportNames}
\alias{reportNames}
\alias{reportNames,AssociationResults-method}
\alias{traitNames,AssociationResults-method}
\alias{reportNames,LDResults-method}
\alias{reportNames,MDSResults-method}
\alias{reportNames,PCAResults-method}
\title{Return report names}
\usage{
reportNames(object)

\S4method{reportNames}{AssociationResults}(object)

\S4method{traitNames}{AssociationResults}(object)

\S4method{reportNames}{LDResults}(object)

\S4method{reportNames}{MDSResults}(object)

\S4method{reportNames}{PCAResults}(object)
}
\arguments{
\item{object}{an \code{\linkS4class{AssociationResults}},
\code{\linkS4class{PCAResults}}, \code{\linkS4class{MDSResults}}, or
\code{\linkS4class{LDResults}} object}
}
\value{
A \code{character} vector of report names.
}
\description{
Returns a \code{character} vector of table report names
}
