% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_tassel_distance_matrix.R
\name{rownames,TasselDistanceMatrix-method}
\alias{rownames,TasselDistanceMatrix-method}
\title{Row names}
\usage{
\S4method{rownames}{TasselDistanceMatrix}(x)
}
\arguments{
\item{x}{An object of class \code{TasselDistanceMatrix}.}
}
\description{
Get row names of a \code{TasselDistanceMatrix} object.
}
