% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rTASSEL-package.R
\docType{package}
\name{rTASSEL-package}
\alias{rTASSEL}
\alias{rTASSEL-package}
\title{rTASSEL: R Front-End for TASSEL}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The aim of rTASSEL is to provide an R front-end for the Java program, TASSEL. TASSEL is a software package used to evaluate trait associations, evolutionary patterns, and linkage disequilibrium. Additionally, this software allows for a number of powerful statistical approaches to association mapping such general and mixed linear modelling.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/maize-genetics/rTASSEL} \url{https://rtassel.maizegenetics.net/}
}

}
\author{
\strong{Maintainer}: Brandon Monier \email{brandon.monier@gmail.com} (\href{https://orcid.org/0000-0001-6797-1221}{ORCID})

Authors:
\itemize{
  \item Brandon Monier \email{brandon.monier@gmail.com} (\href{https://orcid.org/0000-0001-6797-1221}{ORCID})
  \item Terry Casstevens (\href{https://orcid.org/0000-0001-7602-0487}{ORCID})
  \item Peter Bradbury (\href{https://orcid.org/0000-0003-3825-8480}{ORCID})
  \item Ed Buckler (\href{https://orcid.org/0000-0002-3100-371X}{ORCID})
}

}
\keyword{internal}
