% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PullFunctions.R
\name{sampleDataFrame}
\alias{sampleDataFrame}
\title{Get data frame of sample taxa from GenotypeTable}
\usage{
sampleDataFrame(jtsGenoTableOrTaxaList)
}
\arguments{
\item{jtsGenoTableOrTaxaList}{a \code{GenotypeTable} or \code{TaxaList}
class object}
}
\description{
Extracts a data frame of sample taxa from
   \code{GenotypeTable} or \code{TaxaList} class objects.
}
