% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_diversity_functions.R
\name{seqDiversity}
\alias{seqDiversity}
\title{Calculate sequence diversity}
\usage{
seqDiversity(
  tasObj,
  startSite = 0,
  endSite = NULL,
  slidingWindowAnalysis = FALSE,
  stepSize = 100,
  windowSize = 500
)
}
\arguments{
\item{tasObj}{An object of class \code{\linkS4class{TasselGenotype}} or
\code{\linkS4class{TasselGenomicDataset}}. Objects of the deprecated
\code{TasselGenotypePhenotype} class are still accepted.}

\item{startSite}{Start site.}

\item{endSite}{End site. Defaults to the maximum index of markers.}

\item{slidingWindowAnalysis}{Do you want to analyze diversity in a sliding
window? Defaults to \code{FALSE}.}

\item{stepSize}{Step size.}

\item{windowSize}{Window size.}
}
\description{
This performs basic diversity analyses on genetic data,
  specifically, average pairwise divergence (\eqn{\pi}), segregating sites,
  estimated mutation rate (\eqn{\theta}, or \eqn{4N\mu}), and Tajima's D can
  be calculated, as well as sliding windows of diversity.
}
