% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setActiveTASSELVersion}
\alias{setActiveTASSELVersion}
\title{Record the TASSEL version that rTASSEL should load}
\usage{
setActiveTASSELVersion(version)
}
\arguments{
\item{version}{TASSEL version string.}
}
\value{
\code{version}, invisibly.
}
\description{
Record the TASSEL version that rTASSEL should load
}
\keyword{internal}
