% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{methods}
\name{show,TasselGenotypePhenotype-method}
\alias{show,TasselGenotypePhenotype-method}
\title{Show method TasselGenotypePhenotype objects}
\usage{
\S4method{show}{TasselGenotypePhenotype}(object)
}
\arguments{
\item{object}{a \code{TasselGenotypePhenotype} class object}
}
\description{
Prints out information related taxa, positions, genotype, and
   phenotype information.
}
