% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_tassel_phenotype.R
\name{show,TasselPhenotype-method}
\alias{show,TasselPhenotype-method}
\title{Display summary information of a TasselPhenotype object}
\usage{
\S4method{show}{TasselPhenotype}(object)
}
\arguments{
\item{object}{A \code{TasselPhenotype} object}
}
\description{
Display summary information of a TasselPhenotype object
}
