% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_association_results.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,}
\alias{AssociationResults-method}
\title{Show methods for AssociationResults objects}
\usage{
\S4method{show}{AssociationResults}(object)
}
\arguments{
\item{object}{a \code{\linkS4class{AssociationResults}} object}
}
\description{
Prints outs summary information from \code{AssociationResults} objects
}
