% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_tassel_genotype.R
\name{siteIds}
\alias{siteIds}
\title{Select Sites by Name}
\usage{
siteIds(...)
}
\arguments{
\item{...}{Character site name strings.}
}
\value{
A \code{\linkS4class{SiteSelector}} object.
}
\description{
Creates a \code{\linkS4class{SiteSelector}} for filtering by
marker or SNP ID strings.
}
\examples{
\dontrun{
gt[, siteIds("rs1", "rs2")]
}

}
