% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging_support.R
\name{startLogger}
\alias{startLogger}
\title{Start TASSEL logging information}
\usage{
startLogger(filePath = NULL, verbose = TRUE)
}
\arguments{
\item{filePath}{File path and name of log file location. If \code{NULL}, a logging file
(\code{rtasel_log.txt} will be added to current working directory.}
}
\description{
This function will create a file for storing logging output
   from TASSEL.
}
