% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{tasselArtifactUrl}
\alias{tasselArtifactUrl}
\title{Build the Maven Central URL for a TASSEL artifact}
\usage{
tasselArtifactUrl(version, filename)
}
\arguments{
\item{version}{TASSEL version string.}

\item{filename}{Artifact filename.}
}
\value{
A character string URL.
}
\description{
Build the Maven Central URL for a TASSEL artifact
}
\keyword{internal}
