% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_tassel_numeric_genotype.R
\docType{class}
\name{TasselNumericGenotype-class}
\alias{TasselNumericGenotype-class}
\alias{show,TasselNumericGenotype-method}
\title{TasselNumericGenotype Class Definition}
\usage{
\S4method{show}{TasselNumericGenotype}(object)
}
\arguments{
\item{object}{An object of class \code{TasselNumericGenotype}.}
}
\description{
Defines the `TasselNumericGenotype` class, which extends the 
`TasselGenotype` class. This class is used to represent numeric 
genotype data in the TASSEL 5 framework.

This method is used to display information about a 
\code{TasselNumericGenotype} object. It prints a summary of the 
genotype data, including the number of taxa, number of sites, and 
memory address of the Java object.
}
\details{
The function \code{printNumGtDisp} is called internally to format 
and display the genotype data. The number of taxa and sites are 
retrieved from the Java reference object associated with the 
\code{TasselNumericGenotype} instance.
}
