% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\docType{methods}
\name{taxa}
\alias{taxa}
\alias{taxa,GenotypeTable-method}
\title{Get taxa from TASSEL objects}
\usage{
taxa(object)

\S4method{taxa}{GenotypeTable}(object)
}
\description{
Get taxa from TASSEL objects

Get taxa for GenotypeTable class objects
}
