% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_genotype_selectors.R
\name{taxa}
\alias{taxa}
\title{Select Taxa by ID}
\usage{
taxa(...)
}
\arguments{
\item{...}{Character taxa IDs to select.}
}
\value{
A \code{\linkS4class{TaxaSelector}} object.
}
\description{
Creates a \code{\linkS4class{TaxaSelector}} for filtering by taxa
names.
}
\details{
On phenotype data a taxon may hold more than one observation, in
which case every observation of a named taxon is kept.
}
\examples{
\dontrun{
gt[taxa("B73", "Mo17"), ]
ph[taxa("B73", "Mo17"), ]
}

}
