% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{TaxaList-class}
\alias{TaxaList-class}
\title{A R Wrapper for the TaxaList class}
\description{
A R Wrapper for the TaxaList class
}
\details{
Class \code{TaxaList} defines a TaxaList from TASSEL
}
