% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_all_generics.R,
%   R/class_tassel_distance_matrix.R, R/class_tassel_phenotype.R,
%   R/class_tassel_genotype.R, R/class_tassel_genomic_dataset.R,
%   R/taxa_list_functions.R
\name{taxaList}
\alias{taxaList}
\alias{taxaList,TasselDistanceMatrix-method}
\alias{taxaList,TasselPhenotype-method}
\alias{taxaList,TasselGenotype-method}
\alias{taxaList,TasselGenomicDataset-method}
\alias{taxaList,TasselGenotypePhenotype-method}
\title{Get taxa IDs from genotype data}
\usage{
taxaList(tasObj)

\S4method{taxaList}{TasselDistanceMatrix}(tasObj)

\S4method{taxaList}{TasselPhenotype}(tasObj)

\S4method{taxaList}{TasselGenotype}(tasObj)

\S4method{taxaList}{TasselGenomicDataset}(tasObj)

\S4method{taxaList}{TasselGenotypePhenotype}(tasObj)
}
\arguments{
\item{tasObj}{A \code{TasselGenotypePhenotype} or \code{TasselGenotype}
object containing genotype data.}
}
\value{
A character vector of taxa IDs.
}
\description{
Returns a character vector of taxa (sample) IDs from a
\code{\linkS4class{TasselGenotypePhenotype}} or
\code{\linkS4class{TasselGenotype}} object.
}
