% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_genotype_selectors.R
\docType{class}
\name{TaxaSelector-class}
\alias{TaxaSelector-class}
\alias{!,TaxaSelector-method}
\title{TaxaSelector Class}
\usage{
\S4method{!}{TaxaSelector}(x)
}
\arguments{
\item{x}{A \code{TaxaSelector} object.}
}
\description{
S4 class representing taxa selection criteria for bracket-based
filtering of \code{TasselGenotype} objects.
}
\section{Slots}{

\describe{
\item{\code{type}}{Character indicating selector type: \code{"ids"} or
\code{"predicate"}.}

\item{\code{ids}}{Character vector of taxa IDs (used when
\code{type = "ids"}).}

\item{\code{quo}}{A quosure, or a list of quosures combined with \code{&},
for predicate evaluation (used when \code{type = "predicate"}).}

\item{\code{negate}}{Logical indicating whether to negate the selection.}
}}

