% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_all_generics.R,
%   R/class_tassel_phenotype.R, R/class_tassel_genomic_dataset.R
\name{traitNames}
\alias{traitNames}
\alias{traitNames,TasselPhenotype-method}
\alias{traitNames,TasselGenomicDataset-method}
\title{Return trait names}
\usage{
traitNames(object)

\S4method{traitNames}{TasselPhenotype}(object)

\S4method{traitNames}{TasselGenomicDataset}(object)
}
\arguments{
\item{object}{a \code{\linkS4class{AssociationResults}} object}
}
\description{
Returns a \code{character} vector of trait names
}
