% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_plot_association.R
\name{traitValidityChecker}
\alias{traitValidityChecker}
\title{Check for valid traits in association results}
\usage{
traitValidityChecker(trait, assocRes)
}
\arguments{
\item{trait}{A \code{character} vector of trait IDs}

\item{assocRes}{An object of type \code{AssociationResults}}
}
\description{
Check for valid traits in association results
}
