% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_phylogeny_vis_functions.R
\name{treeJavaApp}
\alias{treeJavaApp}
\title{R interface for Archaeopteryx interactive tree viewer}
\usage{
treeJavaApp(tasObj, clustMethod = c("Neighbor_Joining", "UPGMA"))
}
\arguments{
\item{tasObj}{An object of class \code{TasselGenotypePenotype}.}

\item{clustMethod}{What clustering method should be used? Current options
are \code{UGMA} and \code{Neighbor_Joining}. Defaults to
\code{Neighbor_Joining}.}
}
\value{
Returns a Java-based visualization application.
}
\description{
This function acts as a wrapper for TASSEL's
   interface to the Archaeopteryx Java tree Viewer.
}
