% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_tassel_distance_matrix.R
\name{as.matrix.TasselDistanceMatrix}
\alias{as.matrix.TasselDistanceMatrix}
\title{Coerce matrix from TasselDistanceMatrix class}
\usage{
\method{as.matrix}{TasselDistanceMatrix}(x, ...)
}
\arguments{
\item{x}{An object of \code{TasselDistanceMatrix} class.}

\item{...}{Additional arguments to be passed to or from methods.}
}
\description{
Coerces an object of class \code{TasselDistanceMatrix} to
   a \code{matrix} object.
}
