% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassTasselDistanceMatrix.R
\name{as.matrix.TasselDistanceMatrix}
\alias{as.matrix.TasselDistanceMatrix}
\title{Coerce matrix from TasselDistanceMatrix class}
\usage{
\method{as.matrix}{TasselDistanceMatrix}(object)
}
\arguments{
\item{object}{An object of \code{TasselDistanceMatrix} class.}
}
\description{
Coerces an object of class \code{TasselDistanceMatrix} to
   a \code{matrix} object.
}
