% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_tassel_genomic_dataset.R
\name{as.matrix.TasselGenomicDataset}
\alias{as.matrix.TasselGenomicDataset}
\title{Coerce a genomic dataset's genotype data to an R matrix}
\usage{
\method{as.matrix}{TasselGenomicDataset}(x, ...)
}
\arguments{
\item{x}{A \code{TasselGenomicDataset} object.}

\item{...}{Additional arguments to be passed to or from methods.}
}
\value{
An \code{integer} matrix of taxa (rows) by sites (columns).
}
\description{
Returns the joined genotype table held by a
\code{\linkS4class{TasselGenomicDataset}} as a matrix of dosage values.
}
