% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_association_results.R
\docType{class}
\name{AssociationResults-class}
\alias{AssociationResults-class}
\title{AssociationResults Class}
\description{
Class \code{AssociationResults} defines a \code{rTASSEL}
Class for storing TASSEL 5 and general-purpose association results
from *WAS studies.
}
\section{Slots}{

\describe{
\item{\code{results}}{A list of \code{data.frame} objects containing summary results}

\item{\code{traits}}{A vector of type \code{character} containing the trait IDs
modeled.}
}}

