% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_association_results.R
\docType{class}
\name{AssociationResults-class}
\alias{AssociationResults-class}
\alias{show,AssociationResults-method}
\title{AssociationResults Class}
\usage{
\S4method{show}{AssociationResults}(object)
}
\arguments{
\item{object}{a \code{\linkS4class{AssociationResults}} object}
}
\description{
Class \code{AssociationResults} defines a \code{rTASSEL}
Class for storing TASSEL 5 and general-purpose association results
from *WAS studies.

Prints outs summary information from \code{AssociationResults} objects
}
\section{Slots}{

\describe{
\item{\code{results}}{A list of \code{data.frame} objects containing summary results}

\item{\code{traits}}{A vector of type \code{character} containing the trait IDs
modeled.}

\item{\code{assocType}}{\code{character} object describing association type}
}}

