% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_association_results_sub.R
\docType{class}
\name{AssociationResultsBLUE-class}
\alias{AssociationResultsBLUE-class}
\title{AssociationResultsBLUE Class}
\description{
Class \code{AssociationResultsBLUE} defines a \code{rTASSEL}
Class for storing TASSEL 5 BLUE GWAS results
}
