% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PluginSupport.R
\name{assocModelDesign}
\alias{assocModelDesign}
\title{Converts a formula into a GWAS model design that can be used by
   TASSEL PhenotypeGenotype}
\usage{
assocModelDesign(phenotypeGenotype, fixed, random = NULL,
  kinship = NULL)
}
\arguments{
\item{phenotypeGenotype}{TASSEL PhenotypeGenotype Object}

\item{kinship}{A TASSEL kinship matrix}

\item{assocFormula}{Using R formula (lm) notation}
}
\value{
tibble dataframe defining data, covariates, factors, taxa, and genotypes to be used
}
\description{
Converts a formula into a GWAS model design that can be used by
   TASSEL PhenotypeGenotype
}
\details{
. is used for all variable, + to add one, - to exclude, list() used for multiple responses
}
