% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PluginSupport.R
\name{assocModelDesign}
\alias{assocModelDesign}
\title{Title Converts a formula into a GWAS model design that can be used by TASSEL PhenotypeGenotype}
\usage{
assocModelDesign(phenotypeGenotype, assocFormula)
}
\arguments{
\item{phenotypeGenotype}{TASSEL PhenotypeGenotype Object}

\item{assocFormula}{Using R formula (lm) notation}
}
\value{
tibble dataframe defining data, covariates, factors, taxa, and genotypes to be used
}
\description{
Title Converts a formula into a GWAS model design that can be used by TASSEL PhenotypeGenotype
}
\details{
. is used for all variable, + to add one, - to exclude, list() used for multiple responses
}
