% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysisAssociationFunctions.R
\name{assocModelFitter}
\alias{assocModelFitter}
\title{R interface for TASSEL's association methods}
\usage{
assocModelFitter(tasObj, formula, fitMarkers = FALSE, kinship = NULL,
  fastAssociation = FALSE, minClassSize = 0, biallelicOnly = FALSE,
  appendAddDom = FALSE)
}
\arguments{
\item{tasObj}{An object of class \code{TasselGenotypePenotype}.}

\item{formula}{An R-based linear model formula. The general layout of this
formula uses the following TASSEL data scheme:
\code{<data> ~ <factor> and/or <covariate>}. If all traits in a Phenotype
object should be ran, a simplified formula (\code{. ~ .}) can be used.
This scheme can also be used for running all \code{<data>} or
\code{<factor>} and/or \code{<covariate>} data as well. Single variables
are separated witha \code{+} operator. See vignette for further
clarification.}

\item{fitMarkers}{Should marker data be fitted? If \code{TRUE}, GLM
analysis will be executed. If \code{FALSE}, BLUEs will be calculated.
Defaults to \code{FALSE}.}

\item{kinship}{Should kinship data be accounted for in the model? If so,
please submit a TASSEL kinship matrix object using the rTASSEL function,
\code{kinshipMatrix}. Defaults to \code{NULL}}

\item{fastAssociation}{Should TASSEL's Fast Association plugin be used?
Consider setting to \code{TRUE} if you have many phenotypes in your
data set.}

\item{minClassSize}{The minimum acceptable genotype class size. Genotypes
in a class with a smaller size will be set to missing. Defaults to 0.}

\item{biallelicOnly}{Only test sites that are bi-allelic. The alternative is
to test sites with two or more alleles. Defaults to \code{FALSE}}

\item{appendAddDom}{If true, additive and dominance effect estimates will
be added to the stats report for bi-allelic sites only. The effect will
only be estimated when the data source is genotype (not a probability).
The additive effect will always be non-negative. Defaults to \code{FALSE}}
}
\value{
Returns an R list containing \code{tibble}-based data frames
}
\description{
This function acts as a front-end for TASSEL's extensive
   association analysis methods. Using this function, users can run
   the following TASSEL association methods:
   \itemize{
     \item{best linear unbiased estimates (BLUEs)}
     \item{generalized linear model (GLM)}
     \item{mixed linear model}
     \item{Fast association (Shabalin 2012)}
   }
}
