% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_all_generics.R,
%   R/class_tassel_phenotype.R
\name{attributeData}
\alias{attributeData}
\alias{attributeData,TasselPhenotype-method}
\title{Return TASSEL attribute data}
\usage{
attributeData(object, ...)

\S4method{attributeData}{TasselPhenotype}(object)
}
\arguments{
\item{object}{an \code{rTASSEL} object}

\item{...}{Additional arguments, for use in specific methods}
}
\description{
Returns attribute data from an rTASSEL phenotype object
}
