% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_methods.R
\name{concatenate}
\alias{concatenate}
\title{Concatenate phenotype tables}
\usage{
concatenate(x)
}
\arguments{
\item{x}{A vector of phenotype objects.}
}
\description{
Concatenate (e.g. bind rows) multiple phenotype objects based
   on \code{Taxa} column.
}
