% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_phylogeny_functions.R
\name{createTree}
\alias{createTree}
\title{R interface for TASSEL's tree creation methods}
\usage{
createTree(tasObj, clustMethod = c("Neighbor_Joining", "UPGMA"))
}
\arguments{
\item{tasObj}{An object of class \code{TasselGenotypePenotype}.}

\item{clustMethod}{What clustering method should be used? Current options
are \code{UGMA} and \code{Neighbor_Joining}. Defaults to
\code{Neighbor_Joining}.}
}
\value{
Returns a \code{phylo} tree object. See the
   [ape]{https://cran.r-project.org/web/packages/ape/ape.pdf} package
   for further details.
}
\description{
This function acts as a wrapper for TASSEL's
   \code{CreateTreePlugin}.
}
