% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysisRelatednessFunctions.R
\name{distanceMatrix}
\alias{distanceMatrix}
\title{Create a TASSEL distance matrix}
\usage{
distanceMatrix(tasObj)
}
\arguments{
\item{tasObj}{an rTASSEL \code{TasselGenotypePhenotype} object}
}
\value{
Returns a Java pointer of a TASSEL distance matrix object.
}
\description{
This function will calculate a distance matrix using
   TASSEL's \code{DistanceMatrix} plugin.
}
