% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportUtils.R
\name{exportGenotypeTable}
\alias{exportGenotypeTable}
\title{Export Genotype Table to Disk}
\usage{
exportGenotypeTable(
  tasObj,
  file = "",
  format = c("vcf", "hapmap", "plink", "flapjack")
)
}
\arguments{
\item{tasObj}{An object of class \code{TasselGenotypePenotype} that
contains a genotype table.}

\item{file}{Output file name.}

\item{format}{Export file format.}
}
\description{
Exports genotype tables to various flat file formats.
}
