% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_functions.R
\name{filterGenotypeTableBySiteName}
\alias{filterGenotypeTableBySiteName}
\title{Filter genotype table by site IDs}
\usage{
filterGenotypeTableBySiteName(tasObj, siteNames)
}
\arguments{
\item{tasObj}{An object of class \code{TasselGenotypePenotype}.}

\item{siteNames}{A character vector of site names to filter on.}
}
\value{
Returns an object of \code{TasselGenotypePhenotype} class.
}
\description{
Filter a genotype table object by specifying literal site names (IDs)
for variant markers.
}
