% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterFunctions.R
\name{filterGenotypeTableSites}
\alias{filterGenotypeTableSites}
\title{Filter genotype table by sites}
\usage{
filterGenotypeTableSites(
  tasObj,
  siteMinCount = 0,
  siteMinAlleleFreq = 0,
  siteMaxAlleleFreq = 1,
  minHeterozygous = 0,
  maxHeterozygous = 1,
  siteRangeFilterType = c("none", "sites", "position"),
  startSite = NULL,
  endSite = NULL,
  startChr = NULL,
  startPos = NULL,
  endChr = NULL,
  endPos = NULL,
  chrPosFile = NULL,
  bedFile = NULL
)
}
\arguments{
\item{tasObj}{An object of class \code{TasselGenotypePenotype}.}

\item{siteMinCount}{Site minimum count of alleles not unknown. Can range
from 0 to inf. Defaults to 0.}

\item{siteMinAlleleFreq}{Site minimum minor allele frequency. Can range
from 0 to 1.0. Defaults to 0.0.}

\item{siteMaxAlleleFreq}{Site maximum minor allele frequency. Can range
from 0 to 1.0. Defaults to 1.0.}

\item{minHeterozygous}{Min heterozygous proportion. Can range from 0 to 1.0.
Defaults to 0.0.}

\item{maxHeterozygous}{Max heterozygous proportion. Can range from 0 to 1.0.
Defaults to 1.0.}

\item{siteRangeFilterType}{True if filtering by site numbers. False if
filtering by chromosome and position. Options are
\code{none}, \code{sites}, or \code{position}. Defaults to \code{NONE}.}

\item{startSite}{The start site. Defaults to 0.}

\item{endSite}{The end site. Defaults to 0.}

\item{startChr}{Start chromosome for site filtration range if \code{position}
is chosen from \code{siteRangeFilterType}. Needs end chromsome
(\code{endChr}) to work.}

\item{startPos}{Physical start position (bp) for filtration range if
\code{position} is chosen from \code{siteRangeFilterType}. If
\code{NULL}, the first physical position in the data set will be
chosen.}

\item{endChr}{End chromosome for site filtration range if \code{position}
is chosen from \code{siteRangeFilterType}. Needs start chromsome
(\code{endChr}) to work.}

\item{endPos}{Physical end position (bp) for filtration range if
\code{position} is chosen from \code{siteRangeFilterType}. If
\code{NULL}, the last physical position in the data set will be
chosen.}

\item{chrPosFile}{An optional chromosome position file path of
\code{character} class. Defaults to \code{NULL}. \strong{Note:}
a chromosome position file must contain correct formatting
(e.g. a two column file with the header of
\code{c("Chromosome", "Position")}).}

\item{bedFile}{An optional BED coordinate file path of
\code{character} class. Defaults to \code{NULL}.}
}
\value{
Returns an object of \code{TasselGenotypePhenotype} class.
}
\description{
This function will filter R objects of
   \code{TasselGenotypePhenotype} class containing genotype tables.
   The parameters for this function are derived from TASSEL's
   \code{FilterSiteBuilder} plugin.
}
