% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterFunctions.R
\name{filterGenotypeTableSites}
\alias{filterGenotypeTableSites}
\title{Filter genotype table by sites}
\usage{
filterGenotypeTableSites(tasObj, siteMinCount = 0,
  siteMinAlleleFreq = 0, siteMaxAlleleFreq = 1, minHeterozygous = 0,
  maxHeterozygous = 1, siteRangeFilterType = "NONE", startSite = 0,
  endSite = 0)
}
\arguments{
\item{tasObj}{An object of class \code{TasselGenotypePenotype}.}

\item{siteMinCount}{Site minimum count of alleles not unknown. Can range
from 0 to inf. Defaults to 0.}

\item{siteMinAlleleFreq}{Site minimum minor allele frequency. Can range
from 0 to 1.0. Defaults to 0.0.}

\item{siteMaxAlleleFreq}{Site maximum minor allele frequency. Can range
from 0 to 1.0. Defaults to 1.0.}

\item{minHeterozygous}{Min heterozygous proportion. Can range from 0 to 1.0.
Defaults to 0.0.}

\item{maxHeterozygous}{Max heterozygous proportion. Can range from 0 to 1.0.
Defaults to 1.0.}

\item{siteRangeFilterType}{True if filtering by site numbers. False if
filtering by chromosome and position. Options are
\code{NONE}, \code{SITES}, or \code{POSITIONS}. Defaults to \code{NONE}.}

\item{startSite}{The start site. Defaults to 0.}

\item{endSite}{The end site. Defaults to 0.}
}
\value{
Returns an object of \code{TasselGenotypePhenotype} class.
}
\description{
This function will filter R objects of
   \code{TasselGenotypePhenotype} class containing genotype tables.
   The parameters for this function are derived from TASSEL's
   \code{FilterSiteBuilder} plugin.
}
