% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterFunctions.R
\name{filterGenotypeTableTaxa}
\alias{filterGenotypeTableTaxa}
\title{Filter genotype table by taxa}
\usage{
filterGenotypeTableTaxa(tasObj, minNotMissing = 0, minHeterozygous = 0,
  maxHeterozygous = 1)
}
\arguments{
\item{tasObj}{An object of class \code{TasselGenotypePenotype}.}

\item{minNotMissing}{Minimum proportion of sites not unknown to pass this
filter. Value can be between 0.0 and 1.0.}

\item{minHeterozygous}{Minimum proportion of sites that are heterozygous.
Value can be between 0.0 and 1.0.}

\item{maxHeterozygous}{Maximum proportion of sites that are heterozygous.
Value can be between 0.0 and 1.0.}
}
\value{
Returns an object of \code{TasselGenotypePhenotype} class.
}
\description{
This function will filter R objects of
   \code{TasselGenotypePhenotype} class containing genotype tables.
   The parameters for this function are derived from TASSEL's
   \code{FilterTaxaBuilder} plugin.
}
