% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TasselPluginWrappers.R
\name{filterSiteBuilderPlugin}
\alias{filterSiteBuilderPlugin}
\title{Filter Site Builder plugin}
\usage{
filterSiteBuilderPlugin(genotypeTable, siteMinCount = 0,
  siteMinAlleleFreq = 0, siteMaxAlleleFreq = 1, minHeterozygous = 0,
  maxHeterozygous = 1, startSite = 0, endSite = 0)
}
\description{
Filter Site Builder plugin
}
