% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PullFunctions.R
\name{genomicRanges}
\alias{genomicRanges}
\title{Constructor for GRanges (GenomicRanges) class object}
\usage{
genomicRanges(jtsGenoTable)
}
\arguments{
\item{jtsGenoTable}{a \code{GenotypeTable} class object}
}
\description{
Extracts variant position information from each chromosome
   location and stores it as a GenomicRanges (\code{GRanges}) class object
}
