% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{GenotypeTable}
\alias{GenotypeTable}
\title{Wrapper function for GenotypeTable class object}
\usage{
GenotypeTable(path)
}
\arguments{
\item{path}{a path to a \code{VCF} file}
}
\description{
Wrapper function for GenotypeTable class object
}
