% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhenotypeFunctions.R
\name{getPhenotypeDF}
\alias{getPhenotypeDF}
\title{Get an R/\code{tibble} phenotype data frame from TASSEL object}
\usage{
getPhenotypeDF(tasObj)
}
\arguments{
\item{tasObj}{An object of class \code{TasselGenotypePenotype}.}
}
\value{
Returns an \code{tibble} based data frame
}
\description{
This function will extract a \code{tibble}-based R data
   frame from an object of \code{TasselGenotypePhenotype} class that
   contains phenotypic data. Column data will be converted to the following
   types data depending on TASSEL data type:
   \itemize{
     \item{\code{<taxa>}: \code{character}}
     \item{\code{<data>}: \code{numeric}}
     \item{\code{<covariate>}: \code{numeric}}
     \item{\code{<factor>}: \code{factor}}
   }
}
