% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FactorTableFunctions.R
\name{getSumExp}
\alias{getSumExp}
\title{Create Summarized Experiment from a TASSEL Genotype Table}
\usage{
getSumExp(factorTable)
}
\arguments{
\item{factorTable}{An object of clase \code{FactorTable}.}
}
\value{
Returns a \code{SummarizedExperiment} of TASSEL genotype data.
}
\description{
This function will generate an object of
   \code{SummarizedExperiment} class for marker data derived from a
   \code{FactorTable} class object.
}
