% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenotypeTableFunctions.R
\name{getSumExpFromGenotypeTable}
\alias{getSumExpFromGenotypeTable}
\title{Create Summarized Experiment from a TASSEL Genotype Table}
\usage{
getSumExpFromGenotypeTable(tasObj, useRef = FALSE)
}
\arguments{
\item{tasObj}{An object of class \code{TasselGenotypePenotype}.}

\item{useRef}{Use reference alleles or major alleles at sites. If
\code{FALSE}, major alleles will be used.}
}
\value{
Returns a \code{SummarizedExperiment} of TASSEL genotype data.
}
\description{
This function will generate an object of
   \code{SummarizedExperiment} class for marker data derived from a
   \code{TasselGenotypePhenotype} class object.
}
