% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{getTASSELCacheDir}
\alias{getTASSELCacheDir}
\title{Get the local cache directory for TASSEL JARs}
\usage{
getTASSELCacheDir(version = TASSEL_MAVEN$VERSION)
}
\arguments{
\item{version}{TASSEL version string.}
}
\value{
A character string path to the versioned cache directory.
}
\description{
Get the local cache directory for TASSEL JARs
}
\keyword{internal}
