% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImputationMethods.R
\name{imputeLDKNNi}
\alias{imputeLDKNNi}
\title{LD KNNi imputation}
\usage{
imputeLDKNNi(tasObj, highLDSSites = 30, knnTaxa = 10, maxDistance = 1e+07)
}
\arguments{
\item{tasObj}{an rTASSEL \code{TasselGenotypePhenotype} object.}

\item{highLDSSites}{Number of sites in high LD to use in imputation.
Acceptable values are between \code{2} and \code{2000}. Defaults to
\code{30}.}

\item{knnTaxa}{Number of neighbors to use in imputation. Acceptable values
are between \code{2} and \code{200}. Defaults to \code{10}.}

\item{maxDistance}{Maximum physical distance between sites to search for LD
(-1 for no distance cutoff - unlinked chromosomes will be tested).
Defaults to \code{10e6}.}
}
\description{
This imputation algorithm uses LD to identify good predictors
   for each SNP, and then uses the high LD SNPs to identify K-Nearest
   Neighbors. The genotype is called with a weighted mode of the KNNs.
}
