% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation_methods.R
\name{imputeNumeric}
\alias{imputeNumeric}
\title{Imputation methods in Numerical Transformations}
\usage{
imputeNumeric(
  tasObj,
  byMean = TRUE,
  nearestNeighbors = 5,
  distance = c("Euclidean", "Manhattan", "Cosine")
)
}
\arguments{
\item{tasObj}{an rTASSEL \code{TasselGenotypePhenotype} object.}

\item{byMean}{Will imputation be performed by computing the mean of the
respective column? Defaults to \code{TRUE}.}

\item{nearestNeighbors}{Number of nearest neighbors to be evaluated.
Defaults to \code{5}.}

\item{distance}{Distance type. Options are \code{Euclidean},
\code{Manhattan}, or \code{Cosine}. Defaults to \code{Euclidean}.}
}
\description{
This method takes an input \code{GenotypeTable} object with
   missing values and imputes the missing values using one of the chosen
   methods.
}
