% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_methods.R
\name{intersectJoin}
\alias{intersectJoin}
\title{Intersect join phenotype tables}
\usage{
intersectJoin(x)
}
\arguments{
\item{x}{A list of rTASSEL objects containing a phenotype.}
}
\description{
Intersect join multiple phenotype objects based on \code{Taxa}
   column.
}
