% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysisRelatednessFunctions.R
\name{kinshipMatrix}
\alias{kinshipMatrix}
\title{Create a TASSEL kinship matrix}
\usage{
kinshipMatrix(tasObj, method = "Centered_IBS", maxAlleles = 6,
  algorithmVariation = "Observed_Allele_Freq")
}
\arguments{
\item{tasObj}{An object of class \code{TasselGenotypePenotype}.}

\item{method}{A Kinship method.}

\item{maxAlleles}{Maximum number of alleles.}

\item{algorithmVariation}{Algorithm variation.}
}
\value{
Returns a Java pointer of a TASSEL kinship matrix object.
}
\description{
This function will calculate a kinship matrix using
   TASSEL's \code{Kinship} plugin and respective parameters.
}
