% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysisAssociationFunctions.R
\name{leaveOneFamilyOut}
\alias{leaveOneFamilyOut}
\title{Leave one family out cross-validation}
\usage{
leaveOneFamilyOut(phenoFamilyDF)
}
\arguments{
\item{phenoDF}{An object of class \code{TasselGenotypePenotype} that
contains a phenotype object with family data.}
}
\description{
Cross-validation by leaving one family out.
}
