% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_relatedness_functions.R
\name{mds}
\alias{mds}
\title{Run MDS on \code{TasselDistanceMatrix} objects}
\usage{
mds(distMat, nAxes = 5, removeNaN = TRUE)
}
\arguments{
\item{distMat}{A \code{TasselDistanceMatrix} object.}

\item{nAxes}{The number of axes or dimensions and associated eigenvalues to
be returned by the analysis. Defaults to \code{5}.}

\item{removeNaN}{Remove \code{NaNs} from matrix before performing MDS.
Defaults to \code{TRUE}.}
}
\value{
A \code{tibble} of axes, one row per taxon.
}
\description{
Perform multidimensional scaling (MDS) on
   \code{TasselDistanceMatrix} objects.
}
