% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_methods.R
\name{mergeGenotypeTables}
\alias{mergeGenotypeTables}
\title{Merge genotype tables}
\usage{
mergeGenotypeTables(tasObjL)
}
\arguments{
\item{tasObjL}{A list of objects containing genotype data. Accepted
classes are \code{\linkS4class{TasselGenotype}},
\code{\linkS4class{TasselGenomicDataset}}, and the deprecated
\code{TasselGenotypePhenotype}.}
}
\value{
A \code{\linkS4class{TasselGenotype}} object, or a
\code{TasselGenotypePhenotype} if every input was one.
}
\description{
Merges multiple genotype tables together by site information
}
