% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_methods.R
\name{mergeGenotypeTables}
\alias{mergeGenotypeTables}
\title{Merge genotype tables}
\usage{
mergeGenotypeTables(tasObjL)
}
\arguments{
\item{tasObjL}{A list of objects of class \code{TasselGenotypePenotype}.}
}
\value{
Returns an object of \code{TasselGenotypePhenotype} class.
}
\description{
Merges multiple genotype tables together by site information
}
