% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassTasselDistanceMatrix.R
\name{ncol,TasselDistanceMatrix-method}
\alias{ncol,TasselDistanceMatrix-method}
\title{Number of columns}
\usage{
\S4method{ncol}{TasselDistanceMatrix}(x)
}
\arguments{
\item{x}{An object of class \code{TasselDistanceMatrix}.}
}
\description{
Get number of columns of a \code{TasselDistanceMatrix} object.
}
