% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_relatedness_functions.R
\name{pca}
\alias{pca}
\title{Run PCA on Genotype Table}
\usage{
pca(
  tasObj,
  useCovariance = TRUE,
  limitBy = c("number_of_components", "min_eigenvalue", "total_variance"),
  nComponents = 5,
  minEigenval = 0,
  totalVar = 0.5,
  reportEigenvalues = TRUE,
  reportEigenvectors = TRUE
)
}
\arguments{
\item{tasObj}{An object of class \code{\linkS4class{TasselGenotype}} or
\code{\linkS4class{TasselGenomicDataset}}. Objects of the deprecated
\code{TasselGenotypePhenotype} class are still accepted.}

\item{useCovariance}{If \code{TRUE}, analysis will do an eigenvalue
decomposition of the covariance matrix. If \code{FALSE}, it will use a
correlation matrix. NOTE: Using the covariance matrix is recommended for
genotypes while the correlation matrix is often used for phenotypes.
Defaults to \code{TRUE}.}

\item{limitBy}{This parameter determines the type of value that will be used
to limit the number of principal components (axes) returned. The possible
choices are \code{number_of_components}, \code{min_eigenvalue},
and \code{total_variance}.}

\item{nComponents}{The analysis will return this many principal components
up to the number of taxa.}

\item{minEigenval}{All principal components with an eigenvalue greater than
or equal to this value will be returned. NOTE: works only if
\code{min_eigenvalue} is set in the \code{limitBy} parameter.}

\item{totalVar}{The first principal components that together explain this
proportion of the total variance will be returned. NOTE: works only if
\code{total_variance} is set in the \code{limitBy} parameter.}

\item{reportEigenvalues}{Returns a list of eigenvalues sorted high to low.}

\item{reportEigenvectors}{Returns the eigenvectors calculated from a
Singular Value Decomposition of the data. The resulting table can be
quite large if the number of variants and taxa are big.}
}
\value{
A \code{DataFrame} object.
}
\description{
This method performs principal components analysis and returns
   the requested number of PC axes (components).
}
