% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_pca_results.R
\docType{class}
\name{PCAResults-class}
\alias{PCAResults-class}
\alias{show,PCAResults-method}
\title{PCAResults Class}
\usage{
\S4method{show}{PCAResults}(object)
}
\arguments{
\item{object}{a \code{\linkS4class{PCAResults}} object}
}
\description{
Class \code{PCAResults} defines a \code{rTASSEL}
Class for storing TASSEL 5 PCA results

Prints outs summary information from \code{PCAResults} objects
}
\section{Slots}{

\describe{
\item{\code{results}}{A list of \code{data.frame} objects containing summary results}

\item{\code{jObj}}{An rJava reference object pointing to PCA results in Java memory}
}}

