% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_manhattan.R
\name{plotManhattan}
\alias{plotManhattan}
\title{Create Manhattan plot from rTASSEL association output}
\usage{
plotManhattan(
  assocRes,
  trait = NULL,
  threshold = NULL,
  colors = c("#91baff", "#3e619b"),
  pltTheme = c("default", "classic"),
  showSigMarkers = FALSE
)
}
\arguments{
\item{assocRes}{An object of type \code{AssociationResults}}
}
\description{
This function allows for quick generation of a Manhattan
   plot from rTASSEL association statistical output data.
}
