% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_manhattan_qc.R
\name{plotManhattanQC}
\alias{plotManhattanQC}
\title{Create a QC Manhattan plots from rTASSEL association output}
\usage{
plotManhattanQC(
  assocRes,
  trait = NULL,
  gr,
  window = 1e+05,
  threshold = NULL,
  classicNames = FALSE,
  interactive = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{assocRes}{An object of type \code{AssociationResults}}

\item{trait}{Which phenotypic trait do you want to plot? If set to
\code{NULL}, this will generate a faceted plot with all mapped traits.}

\item{gr}{Genomic ranges of interest. Can be passed as a \code{GRanges}
object or a \code{data.frame} object.}

\item{window}{Window size (base-pairs) flanking surround reference range.
Defaults to \code{100000} (100,000 base-pairs).}

\item{threshold}{User-defined \eqn{-log_{10}(p)}-value threshold for
significant marker determination. Once specified any marker points
higher than this line will be highlighted.}

\item{classicNames}{Do you want to plot classical gene names instead?
NOTE: this will need a \code{classical_id} column for the "ranges of
interest" data. Defaults to \code{FALSE}.}

\item{interactive}{Do you want to produce an interactive visualization?
Defaults to \code{FALSE}.}

\item{verbose}{Should messages be printed to console? Defaults to
\code{FALSE}.}
}
\value{
Returns a \code{ggplot2} object
}
\description{
This function allows for quick generation of a QC
   plot from rTASSEL association statistical output data. The main goal
   of this function is to provide "zoomed" in Manhattan plots that
   typically fall within genomic ranges of interest plus a flanking
   window up- and downstream of the ranges.
}
