% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PositionListFunctions.R
\name{positionList}
\alias{positionList}
\title{Get position list metadata from genotype table}
\usage{
positionList(tasObj)
}
\arguments{
\item{tasObj}{A \code{TasselGenotypePhenotype} object}
}
\description{
Returns positional data from a \code{TasselGenotypePhenotype}
   object
}
