% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{readGenotypeTable}
\alias{readGenotypeTable}
\title{Constructor for GenotypeTable class object}
\usage{
readGenotypeTable(path)
}
\description{
Constructor for GenotypeTable class object
}
