% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenotypeTableFunctions.R
\name{readGenotypeTableFromPath}
\alias{readGenotypeTableFromPath}
\title{Wrapper function of TasselGenotypePhenotype class for genotype
   data}
\usage{
readGenotypeTableFromPath(path, keepDepth = FALSE,
  sortPositions = FALSE)
}
\arguments{
\item{path}{A genotype data path (e.g. \code{*.VCF, *.hmp}, etc.).}

\item{keepDepth}{Should depth be kept? Defaults to \code{FALSE}.}

\item{sortPositions}{Should positions be sorted? Defaults to \code{FALSE}.}
}
\value{
Returns an object of \code{TasselGenotypePhenotype} class.
}
\description{
This function is a wrapper for the
   \code{TasselGenotypePhenotype} class. It is used for storing genotype
   information into a class object.
}
