% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_genotype.R
\name{readNumericGenotypeFromRMatrix}
\alias{readNumericGenotypeFromRMatrix}
\title{Read Genotype Data from R Matrix}
\usage{
readNumericGenotypeFromRMatrix(m, asTGP = TRUE)
}
\arguments{
\item{m}{A numeric matrix where rows represent taxa and columns represent
positions/sites. The matrix values are used to calculate reference
probabilities.}

\item{asTGP}{Should the return object be a "classic" \code{TasselGenotypePhenotype}
object (\code{TRUE}) or should it return a \code{TasselNumericGenotype}
object (\code{FALSE})? Defaults to \code{TRUE}.}
}
\value{
An object of class `TasselNumericGenotype` containing the genotype
data and metadata.
}
\description{
This function constructs a `TasselNumericGenotype` object from an
R matrix by interfacing with the TASSEL Java API. It creates taxa
lists, position lists, reference probabilities, and a genotype
table using the provided matrix data.
}
\details{
The following components are constructed using respective Java builder classes:
\itemize{
  \item Taxa list is created using the \code{TASSEL_JVM$TAXA_LIST_BUILDER}
        Java class.
  \item Position list is constructed using the
        \code{TASSEL_JVM$POSITION_LIST_BUILDER} and
        \code{TASSEL_JVM$GENERAL_POSITION_BUILDER} Java classes.
  \item Reference probabilities are built using the
        \code{TASSEL_JVM$REF_PROBABILITY_BUILDER} Java class.
  \item The genotype table is created using the
        \code{TASSEL_JVM$GENOTYPE_TABLE_BUILDER} Java class.
}
}
