% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_tassel_phenotype.R
\name{readPhenotype}
\alias{readPhenotype}
\title{Read and convert phenotype data into TASSEL 5 phenotype objects}
\usage{
readPhenotype(x, attr = NULL)
}
\arguments{
\item{x}{A character string representing the file path to the phenotype data
or a data frame containing the phenotype data.}

\item{attr}{An optional attribute metadata parameter required when \code{x} is
a data frame. Defaults to \code{NULL}.}
}
\value{
A phenotype object created from the input data.
}
\description{
This function reads phenotype data from either a file path or a
data frame.
}
\details{
\itemize{
  \item
  If \code{x} is a character string, the function assumes it is
  a file path and calls \code{readPhenotypeFromFile(x)}.

  \item
  If \code{x} is a data frame, the function requires the
  \code{attr} parameter to provide metadata and calls
  \code{readPhenotypeFromDf(x, attr)}.

  \item
  If \code{x} is neither a character string nor a data frame,
  the function throws an error.

}
}
\examples{
\dontrun{
# Reading phenotype data from a file
phenotype <- readPhenotype("path/to/phenotype/file.txt")

# Reading phenotype data from a data frame
attrDf <- tibble::tribble(
    ~"col_id",      ~"tassel_attr",
    "taxa_id",      "taxa",
    "plant_height", "data",
    "PC1",          "covariate",
    "yield",        "data",
)
df <- tibble::tribble(
    ~"taxa_id", ~"plant_height", ~"PC1", ~"yield",
    "line_a",   12.3,            0.5,    2,
    "line_b",   22.8,            -1.5,   3,
)

phenotypeDf <- readPhenotype(df, attr = attrDf)
}

}
