% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_tassel_phenotype.R
\name{readPhenotype}
\alias{readPhenotype}
\title{Helper function to build a TasselPhenotype object}
\usage{
readPhenotype(x, attr = NULL)
}
\description{
Helper function to build a TasselPhenotype object
}
