% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhenotypeFunctions.R
\name{readPhenotypeFromDataFrame}
\alias{readPhenotypeFromDataFrame}
\title{Wrapper function of TasselGenotypePhenotype class for phenotype
   data from an R data frame}
\usage{
readPhenotypeFromDataFrame(phenotypeDF, taxaID, attributeTypes = NULL)
}
\arguments{
\item{phenotypeDF}{A \code{R} object of class \code{data.frame}.}

\item{taxaID}{The column name that represents your taxa data as a string.}

\item{attributeTypes}{A vector of non-taxa attributes. If \code{NULL}, all
attributes will be TASSEL \code{<data>} types.}
}
\value{
Returns an object of \code{TasselGenotypePhenotype} class.
}
\description{
This function is a wrapper for the
   \code{TasselGenotypePhenotype} class. It is used for storing phenotype
   information into a class object.
}
