% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhenotypeFunctions.R
\name{readPhenotypeFromPath}
\alias{readPhenotypeFromPath}
\title{Wrapper function of TasselGenotypePhenotype class for phenotype
   data from a path.}
\usage{
readPhenotypeFromPath(path)
}
\arguments{
\item{path}{A phenotype data path.}
}
\value{
Returns an object of \code{TasselGenotypePhenotype} class.
}
\description{
This function is a wrapper for the
   \code{TasselGenotypePhenotype} class. It is used for storing phenotype
   information into a class object. This will read in phenotype data from
   a path.
}
