% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FactorTableFunctions.R
\name{refRanges}
\alias{refRanges}
\title{Get reference ranges}
\usage{
refRanges(factorTable)
}
\arguments{
\item{factorTable}{An object of clase \code{FactorTable}.}
}
\value{
Returns a \code{GRanges} object of reference range data.
}
\description{
Returns reference range intervals from a \code{FactorTable}
   object.
}
