% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_genotype_selectors.R
\name{region}
\alias{region}
\title{Select Sites by Genomic Region}
\usage{
region(chrom, start, end)
}
\arguments{
\item{chrom}{Character chromosome ID.}

\item{start}{Numeric start position in bp.}

\item{end}{Numeric end position in bp.}
}
\value{
A \code{\linkS4class{SiteSelector}} object.
}
\description{
Creates a \code{\linkS4class{SiteSelector}} for filtering by a
chromosomal coordinate range.
}
\examples{
\dontrun{
gt[, region("chr1", 1e6, 2e6)]
}

}
