% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_association_results.R
\name{reportNames}
\alias{reportNames}
\alias{reportNames,AssociationResults-method}
\alias{traitNames,AssociationResults-method}
\title{Return report names}
\usage{
reportNames(object)

\S4method{reportNames}{AssociationResults}(object)

\S4method{traitNames}{AssociationResults}(object)
}
\arguments{
\item{object}{a \code{\linkS4class{AssociationResults}} object}
}
\description{
Returns a \code{character} vector of table report names
}
