% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PullFunctions.R
\name{sampleDataFrame}
\alias{sampleDataFrame}
\title{Constructor for GenotypeTable class object}
\usage{
sampleDataFrame(jtsGenoTableOrTaxaList)
}
\description{
Constructor for GenotypeTable class object
}
