% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_tassel_genotype.R
\name{seqnames,TasselGenotype-method}
\alias{seqnames,TasselGenotype-method}
\title{Chromosome (sequence) IDs from TasselGenotype}
\usage{
\S4method{seqnames}{TasselGenotype}(x)
}
\arguments{
\item{x}{A \code{TasselGenotype} object.}
}
\value{
A character vector of chromosome IDs.
}
\description{
Returns the chromosome (sequence) IDs present in the genotype table,
in the order returned by TASSEL.
}
