% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\docType{methods}
\name{show,GenotypeTable-method}
\alias{show,GenotypeTable-method}
\alias{show,PositionList-method}
\alias{show,TaxaList-method}
\title{Display overview when object is called}
\usage{
\S4method{show}{GenotypeTable}(object)

\S4method{show}{PositionList}(object)

\S4method{show}{TaxaList}(object)
}
\description{
Display overview when object is called

Show positions lists

Show method for TaxaList class objects
}
