% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_tassel_genotype.R
\name{show,TasselGenotype-method}
\alias{show,TasselGenotype-method}
\title{Display TasselGenotype Object}
\usage{
\S4method{show}{TasselGenotype}(object)
}
\arguments{
\item{object}{An object of class `TasselGenotype`.}
}
\description{
This method is used to display a summary of a `TasselGenotype`
object. It prints genotype display information, including the
number of taxa, number of sites, and memory address of the Java
object.
}
\details{
The method utilizes the `printGtDisp` function to format and
display the genotype data. It extracts the necessary information
from the `TasselGenotype` object, including the display data
(`dispData`), the number of taxa and sites from the Java reference
object (`jRefObj`), and the memory address (`jMemAddress`).
}
