% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_genotype_selectors.R
\name{sites}
\alias{sites}
\title{Select Sites by Index}
\usage{
sites(...)
}
\arguments{
\item{...}{Integer site indices (0-based).}
}
\value{
A \code{\linkS4class{SiteSelector}} object.
}
\description{
Creates a \code{\linkS4class{SiteSelector}} for filtering by
0-based TASSEL site indices.
}
\examples{
\dontrun{
gt[, sites(0:999)]
gt[, sites(c(10, 50, 100))]
}

}
