% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_all_generics.R,
%   R/class_tassel_genotype.R, R/class_tassel_genomic_dataset.R,
%   R/genotype_table_functions.R
\name{siteSummary}
\alias{siteSummary}
\alias{siteSummary,TasselGenotype-method}
\alias{siteSummary,TasselGenomicDataset-method}
\alias{siteSummary,TasselGenotypePhenotype-method}
\title{Get site summary of genotype table}
\usage{
siteSummary(tasObj)

\S4method{siteSummary}{TasselGenotype}(tasObj)

\S4method{siteSummary}{TasselGenomicDataset}(tasObj)

\S4method{siteSummary}{TasselGenotypePhenotype}(tasObj)
}
\arguments{
\item{tasObj}{A \code{TasselGenotypePhenotype} or \code{TasselGenotype}
object containing genotype data.}
}
\value{
A \code{data.frame} of per-site summary statistics.
}
\description{
Returns per-site summary statistics (allele frequencies, heterozygosity,
missingness, etc.) from genotype data stored in a
\code{\linkS4class{TasselGenotypePhenotype}} or
\code{\linkS4class{TasselGenotype}} object.
}
