% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenotypeTableFunctions.R
\name{siteSummary}
\alias{siteSummary}
\title{Get site summary of genotype table}
\usage{
siteSummary(tasObj)
}
\arguments{
\item{tasObj}{A \code{TasselGenotypePhenotype} object}
}
\description{
Returns positional data from a \code{TasselGenotypePhenotype}
   object
}
