% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoggingSupport.R
\name{startLogger}
\alias{startLogger}
\title{Start TASSEL logging information}
\usage{
startLogger(fullPath = NULL, fileName = NULL, ...)
}
\arguments{
\item{fullPath}{full working path of log file location. If \code{NULL},
logging file will be added to current working directory.}

\item{fileName}{name of logging file. If \code{NULL}, filename will resort
to "rTASSEL_log".}

\item{...}{additional parameters to be added}
}
\description{
This function will create a file for storing logging output
   from TASSEL.
}
