% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_all_generics.R,
%   R/class_association_results.R, R/class_pca_results.R
\name{tableReport}
\alias{tableReport}
\alias{tableReport,AssociationResults-method}
\alias{tableReport,PCAResults-method}
\title{Return selected table report}
\usage{
tableReport(assocRes = missing(), reportName = missing())

\S4method{tableReport}{AssociationResults}(assocRes = missing(), reportName = missing())

\S4method{tableReport}{PCAResults}(assocRes = missing(), reportName = missing())
}
\arguments{
\item{assocRes}{a \code{\linkS4class{AssociationResults}} object}

\item{reportName}{a specific table report to return}
}
\description{
Returns a \code{data.frame} object of association table reports
}
