% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_tassel_distance_matrix.R
\docType{class}
\name{TasselDistanceMatrix-class}
\alias{TasselDistanceMatrix-class}
\title{TasselDistanceMatrix Class}
\description{
Class \code{TasselDistanceMatrix} defines a \code{rTASSEL}
   Class for storing TASSEL genotype and phenotype objects.
}
