% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_tassel_distance_matrix.R
\docType{class}
\name{TasselDistanceMatrix-class}
\alias{TasselDistanceMatrix-class}
\alias{show,TasselDistanceMatrix-method}
\title{TasselDistanceMatrix Class}
\usage{
\S4method{show}{TasselDistanceMatrix}(object)
}
\arguments{
\item{object}{a \code{TasselGenotypePhenotype} class object}
}
\description{
Class \code{TasselDistanceMatrix} defines a \code{rTASSEL}
   Class for storing TASSEL genotype and phenotype objects.

Prints out information related taxa, positions, genotype, and
   phenotype information.
}
