% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_tassel_genotype.R
\docType{class}
\name{TasselGenotype-class}
\alias{TasselGenotype-class}
\title{TasselGenotype Class}
\description{
An S4 class to represent a Tassel Genotype object.
}
\details{
This class is designed to interface with TASSEL 5 for genotype 
data management and analysis. It provides a structure to store and 
interact with Java objects used in TASSEL 5.
}
\section{Slots}{

\describe{
\item{\code{dispData}}{A list containing the display data for the genotype.}

\item{\code{jRefObj}}{A reference to a Java object (`jobjRef`) associated with the genotype.}

\item{\code{jMemAddress}}{A character string representing the memory address of the Java object.}

\item{\code{jClass}}{A character string representing the Java class of the object.}
}}

