% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_tassel_dataset.R
\docType{class}
\name{TasselGenotypePhenotype-class}
\alias{TasselGenotypePhenotype-class}
\alias{show,TasselGenotypePhenotype-method}
\title{TasselGenotypePhenotype Class}
\usage{
\S4method{show}{TasselGenotypePhenotype}(object)
}
\arguments{
\item{object}{a \code{TasselGenotypePhenotype} class object}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Class \code{TasselGenotypePhenotype} defines a \code{rTASSEL}
   Class for storing TASSEL genotype and phenotype objects. It is
   superseded by \code{\linkS4class{TasselGenotype}},
   \code{\linkS4class{TasselPhenotype}}, and
   \code{\linkS4class{TasselGenomicDataset}}, and is scheduled for removal
   in the next major release.

Prints out information related taxa, positions, genotype, and
   phenotype information.
}
