% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{TasselGenotypePhenotype-class}
\alias{TasselGenotypePhenotype-class}
\title{TasselGenotypePhenotype Class}
\description{
Class \code{TasselGenotypePhenotype} defines a \code{rTASSEL}
   Class for storing TASSEL genotype and phenotype objects.
}
