% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_tassel_phenotype.R
\docType{class}
\name{TasselPhenotype-class}
\alias{TasselPhenotype-class}
\title{TasselPhenotype Class Definition}
\description{
Defines the \code{TasselPhenotype} class, which represents
phenotype data in the TASSEL 5 framework.
}
\section{Slots}{

\describe{
\item{\code{attrData}}{A \code{tbl_df} containing attribute data for the phenotype.}

\item{\code{attrSummary}}{A \code{list} summarizing the attributes of the phenotype data.}

\item{\code{dispData}}{A \code{java_pheno_tbl} object for displaying phenotype data.}

\item{\code{rData}}{A \code{tbl_df} containing the phenotype data in R format.}

\item{\code{jRefObj}}{A \code{jobjRef} representing a reference to the Java object in
TASSEL 5.}

\item{\code{jMemAddress}}{A \code{character} string representing the memory address of the
Java object.}

\item{\code{jClass}}{A \code{character} string representing the Java class name of the
object.}
}}

