% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_tassel_genotype.R
\name{taxa}
\alias{taxa}
\title{Select Taxa by ID}
\usage{
taxa(...)
}
\arguments{
\item{...}{Character taxa IDs to select.}
}
\value{
A \code{\linkS4class{TaxaSelector}} object.
}
\description{
Creates a \code{\linkS4class{TaxaSelector}} for filtering a
\code{TasselGenotype} by taxa names.
}
\examples{
\dontrun{
gt[taxa("B73", "Mo17"), ]
}

}
