% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FactorTableFunctions.R
\name{taxa}
\alias{taxa}
\title{Get taxa}
\usage{
taxa(factorTable)
}
\arguments{
\item{factorTable}{An object of clase \code{FactorTable}.}
}
\value{
Returns a vector of taxa of type \code{character}.
}
\description{
Returns taxa IDs from a \code{FactorTable} object.
}
