% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureTableFunctions.R
\name{taxa}
\alias{taxa}
\title{Get taxa}
\usage{
taxa(featureTable)
}
\arguments{
\item{featureTable}{An object of clase \code{FeatureTable}.}
}
\value{
Returns a vector of taxa of type \code{character}.
}
\description{
Returns taxa IDs from a \code{FeatureTable} object.
}
