% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllClasses.R
\docType{methods}
\name{taxa}
\alias{taxa}
\title{Get taxa for GenotypeTable class objects}
\usage{
taxa(object)

\S4method{taxa}{GenotypeTable}(object)
}
\description{
Get taxa for GenotypeTable class objects
}
