% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaxaListFunctions.R
\name{taxaList}
\alias{taxaList}
\title{Get list of taxa from TASSEL data}
\usage{
taxaList(tasObj)
}
\arguments{
\item{tasObj}{A \code{TasselGenotypePhenotype} object}
}
\description{
Returns a list of taxa from a \code{TasselGenotypePhenotype}
   object
}
