% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_all_generics.R,
%   R/class_tassel_genotype.R, R/genotype_table_functions.R
\name{taxaSummary}
\alias{taxaSummary}
\alias{taxaSummary,TasselGenotype-method}
\alias{taxaSummary,TasselGenotypePhenotype-method}
\title{Get taxa summary of genotype table}
\usage{
taxaSummary(tasObj)

\S4method{taxaSummary}{TasselGenotype}(tasObj)

\S4method{taxaSummary}{TasselGenotypePhenotype}(tasObj)
}
\arguments{
\item{tasObj}{A \code{TasselGenotypePhenotype} or \code{TasselGenotype}
object containing genotype data.}
}
\value{
A \code{data.frame} of per-taxon summary statistics.
}
\description{
Returns per-taxon summary statistics (missingness, heterozygosity, etc.)
from genotype data stored in a
\code{\linkS4class{TasselGenotypePhenotype}} or
\code{\linkS4class{TasselGenotype}} object.
}
