% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenotypeTableFunctions.R
\name{taxaSummary}
\alias{taxaSummary}
\title{Get taxa summary of genotype table}
\usage{
taxaSummary(tasObj)
}
\arguments{
\item{tasObj}{A \code{TasselGenotypePhenotype} object}
}
\description{
Returns taxa data from a \code{TasselGenotypePhenotype}
   object
}
