% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_tassel_genotype.R
\name{taxaWhere}
\alias{taxaWhere}
\title{Select Taxa by Predicate}
\usage{
taxaWhere(expr)
}
\arguments{
\item{expr}{An unquoted expression evaluated against a data frame
containing a \code{taxaId} column.}
}
\value{
A \code{\linkS4class{TaxaSelector}} object.
}
\description{
Creates a \code{\linkS4class{TaxaSelector}} using a predicate
expression evaluated against taxa metadata.
}
\examples{
\dontrun{
gt[taxaWhere(startsWith(taxaId, "NAM")), ]
}

}
