% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_association_results.R
\name{traitNames}
\alias{traitNames}
\title{Return trait names}
\usage{
traitNames(object)
}
\arguments{
\item{object}{a \code{\linkS4class{AssociationResults}} object}
}
\description{
Returns a \code{character} vector of trait names
}
