% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_methods.R
\name{unionJoin}
\alias{unionJoin}
\title{Union join phenotype tables}
\usage{
unionJoin(x)
}
\arguments{
\item{x}{A list of rTASSEL objects containing a phenotype. Accepted
classes are \code{\linkS4class{TasselPhenotype}},
\code{\linkS4class{TasselGenomicDataset}},
\code{\linkS4class{PCAResults}}, and the deprecated
\code{TasselGenotypePhenotype}.}
}
\value{
A \code{\linkS4class{TasselPhenotype}} object, or a
   \code{TasselGenotypePhenotype} if every data-bearing input was one.
}
\description{
Union join multiple phenotype objects based on \code{Taxa}
   column.
}
