% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_methods.R
\name{unionJoin}
\alias{unionJoin}
\title{Union join phenotype tables}
\usage{
unionJoin(x)
}
\arguments{
\item{x}{A vector of phenotype objects.}
}
\description{
Union join multiple phenotype objects based on \code{Taxa}
   column.
}
